@@ -4,7 +4,7 @@ from os import path
 
 # Manually specified, more generic version of the software.
 # See: https://stackoverflow.com/a/49684835
-with open('requirements_common.txt') as f:
+with open('requirements.txt') as f:
     requirements = f.readlines()
 
 # Get README
@@ -17,6 +17,38 @@ path_version = path.join(this_directory, 'ivadomed', 'version.txt')
 with open(path_version) as f:
     version = f.read().strip()
 
+extra_requirements = {
+    'docs': [
+        # pin sphinx to match what RTD uses:
+        # https://github.com/readthedocs/readthedocs.org/blob/ecac31de54bbb2c100f933e86eb22b0f4389ba84/requirements/pip.txt#L16
+        'sphinx',
+        'sphinx_rtd_theme',
+        'sphinx-tabs',
+        'sphinx-toolbox',
+        'sphinx-jsonschema',
+        'pypandoc',
+    ],
+    'tests': [
+        'pytest~=6.2',
+        'pytest-cases~=3.6.8',
+        'pytest-cov',
+        'pytest-ordering~=0.6',
+        'pytest-console-scripts~=1.1',
+        'coverage',
+        'coveralls',
+    ],
+    'contrib': [
+        'pre-commit>=2.10.1',
+        'flake8',
+    ]
+}
+
+extra_requirements['dev'] = [
+    requirements,
+    extra_requirements['docs'],
+    extra_requirements['tests'],
+    extra_requirements['contrib'],
+    ]
 
 setup(
     name='ivadomed',
@@ -32,18 +64,11 @@ setup(
         'Intended Audience :: Developers',
         'Programming Language :: Python :: 3',
     ],
-    python_requires='>=3.6, <3.9',
+    python_requires='>=3.7,<3.11',
     packages=find_packages(exclude=['docs', 'tests']),
     include_package_data=True,
     install_requires=requirements,
-    extras_require={
-        'docs': [  # pin sphinx to match what RTD uses:
-            # https://github.com/readthedocs/readthedocs.org/blob/ecac31de54bbb2c100f933e86eb22b0f4389ba84/requirements/pip.txt#L16
-            'sphinx<2',
-            'sphinx-rtd-theme<0.5',
-        ],
-        'dev': ["pre-commit>=2.10.0"]
-    },
+    extras_require=extra_requirements,
     entry_points={
         'console_scripts': [
             'ivadomed=ivadomed.main:run_main',
@@ -54,6 +79,7 @@ setup(
             'ivadomed_convert_to_onnx=ivadomed.scripts.convert_to_onnx:main',
             'ivadomed_extract_small_dataset=ivadomed.scripts.extract_small_dataset:main',
             'ivadomed_download_data=ivadomed.scripts.download_data:main',
+            'ivadomed_segment_image=ivadomed.scripts.segment_image:main',
             'ivadomed_training_curve=ivadomed.scripts.training_curve:main',
             'ivadomed_visualize_and_compare_testing_models=ivadomed.scripts.visualize_and_compare_testing_models:main'
         ],
