@@ -6,17 +6,16 @@
 
 import sys
 import os
-import json
 import numpy as np
 import matplotlib.pyplot as plt
 from sklearn.externals import joblib
 
 from torchvision import transforms as torch_transforms
+from loguru import logger
 
 from ivadomed.loader.bids_dataset import BidsDataset
 from ivadomed import config_manager as imed_config_manager
-from ivadomed.loader import utils as imed_loader_utils
-from ivadomed import utils as imed_utils
+from ivadomed.loader.slice_filter import SliceFilter
 from ivadomed import transforms as imed_transforms
 
 metadata_type = ['FlipAngle', 'EchoTime', 'RepetitionTime']
@@ -48,7 +47,7 @@ def plot_decision_boundaries(data, model, x_range, metadata_name, fname_out):
         plt.xscale('log')
 
     fig.savefig(fname_out)
-    print('\tSave as: ' + fname_out)
+    logger.info(f"\tSave as: {fname_out}")
 
 
 def run_main(context):
@@ -68,7 +67,7 @@ def run_main(context):
                          contrast_lst=context["contrast_train_validation"]
                          if subset != "test" else context["contrast_test"],
                          transform=no_transform,
-                         slice_filter_fn=imed_loader_utils.SliceFilter())
+                         slice_filter_fn=SliceFilter())
 
         for m in metadata_type:
             if m in metadata_dct:
@@ -84,7 +83,7 @@ def run_main(context):
 
     for m in metadata_type:
         values = [v for s in ['train', 'valid', 'test'] for v in metadata_dct[s][m]]
-        print('\n{}: Min={}, Max={}, Median={}'.format(m, min(values), max(values), np.median(values)))
+        logger.info(f"\n{m}: Min={min(values)}, Max={max(values)}, Median={np.median(values)}")
         plot_decision_boundaries(metadata_dct, cluster_dct[m], metadata_range[m], m, os.path.join(out_dir, m + '.png'))
 
 
