@@ -6,18 +6,19 @@
 
 import os
 from tqdm import tqdm
+from loguru import logger
 
 PATH_SCTTESTING = os.path.join(os.path.expanduser('~'), 'duke', 'sct_testing', 'large')
 
 
 def run_main():
     if not os.path.isdir(PATH_SCTTESTING):
-        print('\nThis folder does not exist: {}'.format(PATH_SCTTESTING))
-        print('Please change the path at the top of this file')
+        logger.warning(f"\nThis folder does not exist: {PATH_SCTTESTING}")
+        logger.warning("Please change the path at the top of this file")
 
     subj_lst = [os.path.join(PATH_SCTTESTING, s, 'anat') for s in os.listdir(PATH_SCTTESTING) if
                 os.path.isdir(os.path.join(PATH_SCTTESTING, s, 'anat'))]
-    print('\n{} subjects found.\n'.format(str(len(subj_lst))))
+    logger.info(f"\n{len(subj_lst)} subjects found.\n")
 
     contrast_lst_lst = []
     for subj_fold in tqdm(subj_lst, desc="Scanning dataset"):
@@ -27,9 +28,9 @@ def run_main():
 
     contrast_lst = [sublst for lst in contrast_lst_lst for sublst in lst]
     contrast_lst_noDuplicate = list(set(contrast_lst))
-    print('\n{} contrasts found.\n'.format(str(len(contrast_lst_noDuplicate))))
+    logger.info(f"\n{len(contrast_lst_noDuplicate)} contrasts found.\n")
 
-    print('["{}"]'.format('", "'.join(contrast_lst_noDuplicate)))
+    logger.info(f"['{', '.join(contrast_lst_noDuplicate)}']")
 
 
 if __name__ == "__main__":
