@@ -1,11 +1,10 @@
-import os
-
 import nibabel as nib
 import numpy as np
 import pandas as pd
 from loguru import logger
 from scipy.ndimage import label, generate_binary_structure
 from tqdm import tqdm
+from pathlib import Path
 
 from ivadomed import inference as imed_inference
 from ivadomed import metrics as imed_metrics
@@ -30,19 +29,19 @@ def evaluate(bids_df, path_output, target_suffix, eval_params):
     Returns:
         pd.Dataframe: results for each image.
     """
-    path_preds = os.path.join(path_output, 'pred_masks')
+    path_preds = Path(path_output, 'pred_masks')
     logger.info('\nRun Evaluation on {}\n'.format(path_preds))
 
     # OUTPUT RESULT FOLDER
-    path_results = os.path.join(path_output, 'results_eval')
-    if not os.path.isdir(path_results):
-        os.makedirs(path_results)
+    path_results = Path(path_output, 'results_eval')
+    if not path_results.is_dir():
+        path_results.mkdir(parents=True)
 
-    # INIT DATA FRAME
-    df_results = pd.DataFrame()
+    # INIT DATA FRAME ROW LIST
+    df_lst = []
 
     # LIST PREDS
-    subj_acq_lst = [f.split('_pred')[0] for f in os.listdir(path_preds) if f.endswith('_pred.nii.gz')]
+    subj_acq_lst = [f.name.split('_pred')[0] for f in path_preds.iterdir() if f.name.endswith('_pred.nii.gz')]
 
     # Get all derivatives filenames
     all_deriv = bids_df.get_deriv_fnames()
@@ -50,7 +49,7 @@ def evaluate(bids_df, path_output, target_suffix, eval_params):
     # LOOP ACROSS PREDS
     for subj_acq in tqdm(subj_acq_lst, desc="Evaluation"):
         # Fnames of pred and ground-truth
-        fname_pred = os.path.join(path_preds, subj_acq + '_pred.nii.gz')
+        fname_pred = path_preds.joinpath(subj_acq + '_pred.nii.gz')
         derivatives = bids_df.df[bids_df.df['filename']
                           .str.contains('|'.join(bids_df.get_derivatives(subj_acq, all_deriv)))]['path'].to_list()
         # Ordering ground-truth the same as target_suffix
@@ -77,7 +76,7 @@ def evaluate(bids_df, path_output, target_suffix, eval_params):
         n_classes = len(fname_gt)
         data_gt = np.zeros((h, w, d, n_classes))
         for idx, file in enumerate(fname_gt):
-            if os.path.exists(file):
+            if Path(file).exists():
                 data_gt[..., idx] = nib.load(file).get_fdata()
             else:
                 data_gt[..., idx] = np.zeros((h, w, d), dtype='u1')
@@ -87,25 +86,34 @@ def evaluate(bids_df, path_output, target_suffix, eval_params):
                                    params=eval_params)
         results_pred, data_painted = eval.run_eval()
 
-        # SAVE PAINTED DATA, TP FP FN
-        fname_paint = fname_pred.split('.nii.gz')[0] + '_painted.nii.gz'
-        nib_painted = nib.Nifti1Image(data_painted, nib_pred.affine)
-        nib.save(nib_painted, fname_paint)
+        if eval_params['object_detection_metrics']:
+            # SAVE PAINTED DATA, TP FP FN
+            fname_paint = str(fname_pred).split('.nii.gz')[0] + '_TP-FP-FN.nii.gz'
+            nib_painted = nib.Nifti1Image(
+                dataobj=data_painted.astype(int),
+                affine=nib_pred.header.get_best_affine(),
+                header=nib_pred.header.copy()
+            )
+            nib.save(nib_painted, fname_paint)
 
-        # For Microscopy PNG/TIF files (TODO: implement OMETIFF behavior)
-        if "nii" not in extension:
-            painted_list = imed_inference.split_classes(nib_painted)
-            imed_inference.pred_to_png(painted_list,
-                                       target_suffix,
-                                       os.path.join(path_preds, subj_acq),
-                                       suffix="_painted")
+            # For Microscopy PNG/TIF files (TODO: implement OMETIFF behavior)
+            if "nii" not in extension:
+                painted_list = imed_inference.split_classes(nib_painted)
+                # Reformat target list to include class index and be compatible with multiple raters
+                target_list = ["_class-%d" % i for i in range(len(target_suffix))]
+                imed_inference.pred_to_png(painted_list,
+                                           target_list,
+                                           str(path_preds.joinpath(subj_acq)),
+                                           suffix="_pred_TP-FP-FN.png",
+                                           max_value=3) # painted data contain 3 float values [0.0, 1.0, 2.0, 3.0] corresponding to background, TP, FP and FN objects)
 
         # SAVE RESULTS FOR THIS PRED
         results_pred['image_id'] = subj_acq
-        df_results = df_results.append(results_pred, ignore_index=True)
+        df_lst.append(results_pred)
 
+    df_results = pd.DataFrame(df_lst)
     df_results = df_results.set_index('image_id')
-    df_results.to_csv(os.path.join(path_results, 'evaluation_3Dmetrics.csv'))
+    df_results.to_csv(str(path_results.joinpath('evaluation_3Dmetrics.csv')))
 
     logger.info(df_results.head(5))
     return df_results
@@ -130,6 +138,8 @@ class Evaluation3DMetrics(object):
         bin_struct (ndarray): Binary structure.
         size_min (int): Minimum size of objects. Objects that are smaller than this limit can be removed if
             "removeSmall" is in params.
+        object_detection_metrics (bool): Indicate if object detection metrics (lesions true positive and false detection
+            rates) are computed or not.
         overlap_vox (int): A prediction and ground-truth are considered as overlapping if they overlap for at least this
             amount of voxels.
         overlap_ratio (float): A prediction and ground-truth are considered as overlapping if they overlap for at least
@@ -161,6 +171,8 @@ class Evaluation3DMetrics(object):
         self.bin_struct = generate_binary_structure(3, 2)  # 18-connectivity
         self.postprocessing_dict = {}
         self.size_min = 0
+
+        self.object_detection_metrics = params["object_detection_metrics"]
 
         if "target_size" in params:
             self.size_rng_lst, self.size_suffix_lst = \
@@ -259,14 +271,14 @@ class Evaluation3DMetrics(object):
         data_out = np.zeros(data.shape)
 
         for idx in range(1, n + 1):
-            data_idx = (data_label == idx).astype(np.int)
+            data_idx = (data_label == idx).astype(int)
             n_nonzero = np.count_nonzero(data_idx)
 
             for idx_size, rng in enumerate(self.size_rng_lst):
                 if n_nonzero >= rng[0] and n_nonzero <= rng[1]:
                     data_out[np.nonzero(data_idx)] = idx_size + 1
 
-        return data_out.astype(np.int)
+        return data_out.astype(int)
 
     def get_vol(self, data):
         """Get volume."""
@@ -316,8 +328,8 @@ class Evaluation3DMetrics(object):
         ltp, lfn, n_obj = 0, 0, 0
 
         for idx in range(1, self.n_gt[class_idx] + 1):
-            data_gt_idx = (self.data_gt_label[..., class_idx] == idx).astype(np.int)
-            overlap = (data_gt_idx * self.data_pred).astype(np.int)
+            data_gt_idx = (self.data_gt_label[..., class_idx] == idx).astype(int)
+            overlap = (data_gt_idx * self.data_pred).astype(int)
 
             # if label_size is None, then we look at all object sizes
             # we check if the currrent object belongs to the current size range
@@ -351,11 +363,11 @@ class Evaluation3DMetrics(object):
         """
         lfp = 0
         for idx in range(1, self.n_pred[class_idx] + 1):
-            data_pred_idx = (self.data_pred_label[..., class_idx] == idx).astype(np.int)
-            overlap = (data_pred_idx * self.data_gt).astype(np.int)
+            data_pred_idx = (self.data_pred_label[..., class_idx] == idx).astype(int)
+            overlap = (data_pred_idx * self.data_gt).astype(int)
 
             label_gt = np.max(data_pred_idx * self.data_gt_label[..., class_idx])
-            data_gt_idx = (self.data_gt_label[..., class_idx] == label_gt).astype(np.int)
+            data_gt_idx = (self.data_gt_label[..., class_idx] == label_gt).astype(int)
             # if label_size is None, then we look at all object sizes
             # we check if the current object belongs to the current size range
 
@@ -384,8 +396,12 @@ class Evaluation3DMetrics(object):
             label_size (int): Size of label.
             class_idx (int): Label index. If monolabel 0, else ranges from 0 to number of output channels - 1.
 
-        Note: computed only if n_obj >= 1.
+        Note: computed only if n_obj >= 1 and "object_detection_metrics" evaluation parameter is True.
         """
+        if not self.object_detection_metrics:
+            n_obj = 0
+            return np.nan, n_obj
+
         ltp, lfn, n_obj = self._get_ltp_lfn(label_size, class_idx)
 
         denom = ltp + lfn
@@ -401,8 +417,12 @@ class Evaluation3DMetrics(object):
             label_size (int): Size of label.
             class_idx (int): Label index. If monolabel 0, else ranges from 0 to number of output channels - 1.
 
-        Note: computed only if n_obj >= 1.
+        Note: computed only if n_obj >= 1 and "object_detection_metrics" evaluation parameter is True.
         """
+
+        if not self.object_detection_metrics:
+            return np.nan
+
         ltp, _, n_obj = self._get_ltp_lfn(label_size, class_idx)
         lfp = self._get_lfp(label_size, class_idx)
 
