@@ -1,4 +1,3 @@
-import os
 import nibabel as nib
 import numpy as np
 import onnxruntime
@@ -6,6 +5,7 @@ import torch
 import imageio
 import joblib
 from typing import List
+from pathlib import Path
 
 from loguru import logger
 from torch.utils.data import Dataset
@@ -20,10 +20,12 @@ from ivadomed import models as imed_models
 from ivadomed import postprocessing as imed_postpro
 from ivadomed import transforms as imed_transforms
 from ivadomed.loader import utils as imed_loader_utils, film as imed_film
+from ivadomed.loader.slice_filter import SliceFilter
 from ivadomed.object_detection import utils as imed_obj_detect
 from ivadomed import utils as imed_utils
 from ivadomed import training as imed_training
-
+from ivadomed.keywords import ConfigKW, ModelParamsKW, ObjectDetectionParamsKW, TransformationKW, LoaderParamsKW, \
+    ROIParamsKW, SliceFilterParamsKW, TrainingParamsKW, MetadataKW, OptionKW
 
 
 def onnx_inference(model_path: str, inputs: tensor) -> tensor:
@@ -43,21 +45,20 @@ def onnx_inference(model_path: str, inputs: tensor) -> tensor:
     return torch.tensor(ort_outs[0])
 
 
-def get_preds(context: dict, fname_model: str, model_params: dict, gpu_id: int, batch: dict) -> tensor:
+def get_preds(context: dict, fname_model: str, model_params: dict, cuda_available: bool, device: torch.device, batch: dict) -> tensor:
     """Returns the predictions from the given model.
 
     Args:
         context (dict): configuration dict.
         fname_model (str): name of file containing model.
         model_params (dict): dictionary containing model parameters.
-        gpu_id (int): Number representing gpu number if available. Currently does NOT support multiple GPU segmentation.
+        cuda_available (bool): True if cuda is available.
+        device (torch.device): Device used for prediction.
         batch (dict): dictionary containing input, gt and metadata
 
     Returns:
         tensor: predictions from the model.
     """
-    # Define device
-    cuda_available, device = imed_utils.define_device(gpu_id)
 
     with torch.no_grad():
 
@@ -74,10 +75,10 @@ def get_preds(context: dict, fname_model: str, model_params: dict, gpu_id: int, 
             model.eval()
 
             # Films/Hemis based prediction require meta data load
-            if ('FiLMedUnet' in context and context['FiLMedUnet']['applied']) or \
-                    ('HeMISUnet' in context and context['HeMISUnet']['applied']):
+            if (ConfigKW.FILMED_UNET in context and context[ConfigKW.FILMED_UNET].get(ModelParamsKW.APPLIED)) or \
+                    (ConfigKW.HEMIS_UNET in context and context[ConfigKW.HEMIS_UNET].get(ModelParamsKW.APPLIED)):
                 # Load meta data before prediction
-                metadata = imed_training.get_metadata(batch["input_metadata"], model_params)
+                metadata = imed_training.get_metadata(batch[MetadataKW.INPUT_METADATA], model_params)
                 preds = model(img, metadata)
             else:
                 preds = model(img)
@@ -106,15 +107,18 @@ def get_onehotencoder(context: dict, folder_model: str, options: dict, ds: Datas
     Returns:
         dict: onehotencoder used in the model params.
     """
-    metadata_dict = joblib.load(os.path.join(folder_model, 'metadata_dict.joblib'))
+    metadata_dict = joblib.load(Path(folder_model, 'metadata_dict.joblib'))
     for idx in ds.indexes:
         for i in range(len(idx)):
-            idx[i]['input_metadata'][0][context['FiLMedUnet']['metadata']] = options['metadata']
-            idx[i]['input_metadata'][0]['metadata_dict'] = metadata_dict
+            idx[i][MetadataKW.INPUT_METADATA][0][context[ConfigKW.FILMED_UNET][ModelParamsKW.METADATA]] = options.get(OptionKW.METADATA)
+            idx[i][MetadataKW.INPUT_METADATA][0][MetadataKW.METADATA_DICT] = metadata_dict
 
-    ds = imed_film.normalize_metadata(ds, None, context["debugging"], context['FiLMedUnet']['metadata'])
+    if ConfigKW.DEBUGGING in context and ConfigKW.FILMED_UNET in context and \
+            context[ConfigKW.FILMED_UNET].get(ModelParamsKW.METADATA):
+        ds = imed_film.normalize_metadata(
+            ds, None, context[ConfigKW.DEBUGGING], context[ConfigKW.FILMED_UNET][ModelParamsKW.METADATA])
 
-    return joblib.load(os.path.join(folder_model, 'one_hot_encoder.joblib'))
+    return joblib.load(Path(folder_model, 'one_hot_encoder.joblib'))
 
 
 def pred_to_nib(data_lst: List[np.ndarray], z_lst: List[int], fname_ref: str, fname_out: str, slice_axis: int,
@@ -197,7 +201,7 @@ def pred_to_nib(data_lst: List[np.ndarray], z_lst: List[int], fname_ref: str, fn
     # See: https://github.com/ivadomed/ivadomed/issues/711
     nib_pred = nib.Nifti1Image(
         dataobj=arr_pred_ref_space,
-        affine=None,
+        affine=nib_ref.header.get_best_affine(),
         header=nib_ref.header.copy()
     )
     # save as NifTI file
@@ -207,7 +211,7 @@ def pred_to_nib(data_lst: List[np.ndarray], z_lst: List[int], fname_ref: str, fn
     return nib_pred
 
 
-def pred_to_png(pred_list: list, target_list: list, subj_path: str, suffix: str = ''):
+def pred_to_png(pred_list: list, target_list: list, subj_path: str, suffix: str = '', max_value: int = 1):
     """Save the network predictions as PNG files with suffix "_target_pred".
 
     Args:
@@ -215,18 +219,21 @@ def pred_to_png(pred_list: list, target_list: list, subj_path: str, suffix: str 
         target_list (list of str): list of target suffixes.
         subj_path (str): Path of the subject filename in output folder without extension
             (e.g. "path_output/pred_masks/sub-01_sample-01_SEM").
-        suffix (str): additional suffix to append to the filename.
+        suffix (str): additional suffix to append to the filename (e.g. "_pred.png")
+        max_value (int): Maximum mask value of the float mask to use during the conversion to uint8.
     """
     for pred, target in zip(pred_list, target_list):
-        filename = subj_path + target + "_pred" + suffix + ".png"
+        filename = subj_path + target + suffix
         data = pred.get_fdata()
-        imageio.imwrite(filename, data, format='png')
+        _img = (data*255/max_value).astype(np.uint8).squeeze()
+        imageio.v3.imwrite(filename, _img, extension='.png')
 
 
 def process_transformations(context: dict, fname_roi: str, fname_prior: str, metadata: dict, slice_axis: int,
                             fname_images: list) -> dict:
     """Sets the transformation based on context parameters. When ROI is not provided center-cropping is applied.
-       If there is an object_detection_path, then we modify the metadata to store transformation data.
+
+    If there is an object_detection_path, then we modify the metadata to store transformation data.
 
     Args:
         context (dict): configuration dictionary.
@@ -239,19 +246,22 @@ def process_transformations(context: dict, fname_roi: str, fname_prior: str, met
     Returns:
         dict: metadata.
     """
-    if fname_roi is None and 'ROICrop' in context["transformation"].keys():
+    if fname_roi is None and TransformationKW.ROICROP in context[ConfigKW.TRANSFORMATION].keys():
         logger.warning(
             "fname_roi has not been specified, then a cropping around the center of the image is "
             "performed instead of a cropping around a Region of Interest.")
 
-        context["transformation"] = dict((key, value) if key != 'ROICrop'
-                                         else ('CenterCrop', value)
-                                         for (key, value) in context["transformation"].items())
+        # Convert transformation configuration into dict.
+        for (key, value) in context[ConfigKW.TRANSFORMATION].items():
+            if key != TransformationKW.ROICROP:
+                context[ConfigKW.TRANSFORMATION] = dict(key, value)
+            else:
+                context[ConfigKW.TRANSFORMATION] = dict(TransformationKW.CENTERCROP, value)
 
-    if 'object_detection_params' in context and \
-            context['object_detection_params']['object_detection_path'] is not None:
+    if ConfigKW.OBJECT_DETECTION_PARAMS in context and \
+            context[ConfigKW.OBJECT_DETECTION_PARAMS][ObjectDetectionParamsKW.OBJECT_DETECTION_PATH] is not None:
         imed_obj_detect.bounding_box_prior(fname_prior, metadata, slice_axis,
-                                           context['object_detection_params']['safety_factor'])
+                                           context[ConfigKW.OBJECT_DETECTION_PARAMS][ObjectDetectionParamsKW.SAFETY_FACTOR])
         metadata = [metadata] * len(fname_images)
 
     return metadata
@@ -259,8 +269,9 @@ def process_transformations(context: dict, fname_roi: str, fname_prior: str, met
 
 def set_option(options: dict, postpro: dict, context: dict, key: str):
     """Generalized function that sets postprocessing option based on given list of options.
-       When given key already exists in options, we initialize the key value for the postprocessing dictionary
-       Otherwise, when the key is already found in the postprocessing attritute of the context, we remove it
+
+    When given key already exists in options, we initialize the key value for the postprocessing dictionary
+    Otherwise, when the key is already found in the postprocessing attritute of the context, we remove it
 
     Args:
         options (dict): Contains postprocessing steps information.
@@ -274,8 +285,8 @@ def set_option(options: dict, postpro: dict, context: dict, key: str):
     if options[key]:
         postpro[key] = {}
     # Remove key in context if value set to 0
-    elif key in context['postprocessing']:
-        del context['postprocessing'][key]
+    elif key in context[ConfigKW.POSTPROCESSING]:
+        del context[ConfigKW.POSTPROCESSING][key]
 
 
 def set_postprocessing_options(options: dict, context: dict):
@@ -287,25 +298,25 @@ def set_postprocessing_options(options: dict, context: dict):
     """
     postpro = {}
 
-    if 'binarize_prediction' in options and options['binarize_prediction']:
-        postpro['binarize_prediction'] = {"thr": options['binarize_prediction']}
+    if OptionKW.BINARIZE_PREDICTION in options and options[OptionKW.BINARIZE_PREDICTION] is not None:
+        postpro[OptionKW.BINARIZE_PREDICTION] = {"thr": options[OptionKW.BINARIZE_PREDICTION]}
 
-    if 'binarize_maxpooling' in options and options['binarize_maxpooling'] is not None:
-        set_option(options, postpro, context, 'binarize_maxpooling')
+    if OptionKW.BINARIZE_MAXPOOLING in options and options.get(OptionKW.BINARIZE_MAXPOOLING) is not None:
+        set_option(options, postpro, context, OptionKW.BINARIZE_MAXPOOLING)
 
-    if 'keep_largest' in options and options['keep_largest'] is not None:
-        set_option(options, postpro, context, 'keep_largest')
+    if OptionKW.KEEP_LARGEST in options and options.get(OptionKW.KEEP_LARGEST) is not None:
+        set_option(options, postpro, context, OptionKW.KEEP_LARGEST)
 
-    if 'fill_holes' in options and options['fill_holes'] is not None:
-        set_option(options, postpro, context, 'fill_holes')
+    if OptionKW.FILL_HOLES in options and options.get(OptionKW.FILL_HOLES) is not None:
+        set_option(options, postpro, context, OptionKW.FILL_HOLES)
 
-    if 'remove_small' in options and options['remove_small'] and \
-            ('mm' in options['remove_small'][-1] or 'vox' in options['remove_small'][-1]):
-        unit = 'mm3' if 'mm3' in options['remove_small'][-1] else 'vox'
-        thr = [int(t.replace(unit, "")) for t in options['remove_small']]
-        postpro['remove_small'] = {"unit": unit, "thr": thr}
+    if OptionKW.REMOVE_SMALL in options and options.get(OptionKW.REMOVE_SMALL) and \
+            ('mm' in options[OptionKW.REMOVE_SMALL][-1] or 'vox' in options[OptionKW.REMOVE_SMALL][-1]):
+        unit = 'mm3' if 'mm3' in options[OptionKW.REMOVE_SMALL][-1] else 'vox'
+        thr = [int(t.replace(unit, "")) for t in options[OptionKW.REMOVE_SMALL]]
+        postpro[OptionKW.REMOVE_SMALL] = {"unit": unit, "thr": thr}
 
-    context['postprocessing'].update(postpro)
+    context[ConfigKW.POSTPROCESSING].update(postpro)
 
 
 def segment_volume(folder_model: str, fname_images: list, gpu_id: int = 0, options: dict = None):
@@ -323,31 +334,41 @@ def segment_volume(folder_model: str, fname_images: list, gpu_id: int = 0, optio
             images to segment, e.i., len(fname_images) > 1.
         gpu_id (int): Number representing gpu number if available. Currently does NOT support multiple GPU segmentation.
         options (dict): This can optionally contain any of the following key-value pairs:
-            * 'binarize_prediction': (float) Binarize segmentation with specified threshold.
-                                     Predictions below the threshold become 0, and predictions above or equal to
-                                     threshold become 1. Set to -1 for no thresholding (i.e., soft segmentation).
-            * 'binarize_maxpooling': (bool) Binarize by setting to 1 the voxel having the maximum prediction across
-                                     all classes. Useful for multiclass models.
+
+            * 'binarize_prediction': (float) Binarize segmentation with specified threshold. \
+                Predictions below the threshold become 0, and predictions above or equal to \
+                threshold become 1. Set to -1 for no thresholding (i.e., soft segmentation).
+            * 'binarize_maxpooling': (bool) Binarize by setting to 1 the voxel having the maximum prediction across \
+                all classes. Useful for multiclass models.
             * 'fill_holes': (bool) Fill small holes in the segmentation.
             * 'keep_largest': (bool) Keep the largest connected-object for each class from the output segmentation.
-            * 'remove_small': (list of str) Minimal object size to keep with unit (mm3 or vox). A single value can be provided
-                              or one value per prediction class. Single value example: ["1mm3"], ["5vox"]. Multiple values
-                              example: ["10", "20", "10vox"] (remove objects smaller than 10 voxels for class 1 and 3,
+            * 'remove_small': (list of str) Minimal object size to keep with unit (mm3 or vox). A single value can be provided \
+                              or one value per prediction class. Single value example: ["1mm3"], ["5vox"]. Multiple values \
+                              example: ["10", "20", "10vox"] (remove objects smaller than 10 voxels for class 1 and 3, \
                               and smaller than 20 voxels for class 2).
-            * 'pixel_size': (list of float) List of microscopy pixel size in micrometers.
-                            Length equals 2 [X, Y] for 2D or 3 [X, Y, Z] for 3D.
-            * 'overlap_2D': (list of int) List of overlaps in pixels for 2D patching. Length equals 2 [X, Y].
+            * 'pixel_size': (list of float) List of microscopy pixel size in micrometers. \
+                            Length equals 2 [PixelSizeX, PixelSizeY] for 2D or 3 [PixelSizeX, PixelSizeY, PixelSizeZ] for 3D, \
+                            where X is the width, Y the height and Z the depth of the image.
+            * 'pixel_size_units': (str) Units of pixel size (Must be either "mm", "um" or "nm")
+            * 'no_patch': (bool) 2D patches are not used while segmenting with models trained with patches. \
+                          The "no_patch" option supersedes the "overlap_2D" option. \
+                          This option may not be suitable with large images depending on computer RAM capacity.
+            * 'overlap_2D': (list of int) List of overlaps in pixels for 2D patching. Length equals 2 [OverlapX, OverlapY], \
+                            where X is the width and Y the height of the image.
             * 'metadata': (str) Film metadata.
-            * 'fname_prior': (str) An image filename (e.g., .nii.gz) containing processing information
-                (e.g., spinal cord segmentation, spinal location or MS lesion classification, spinal cord centerline), 
-                used to crop the image prior to segment it if provided.
+            * 'fname_prior': (str) An image filename (e.g., .nii.gz) containing processing information \
+                (e.g., spinal cord segmentation, spinal location or MS lesion classification, spinal cord centerline), \
+                used to crop the image prior to segment it if provided. \
                 The segmentation is not performed on the slices that are empty in this image.
 
     Returns:
-        list: List of nibabel objects containing the soft segmentation(s), one per prediction class.
-        list: List of target suffix associated with each prediction in `pred_list`
+        list, list: List of nibabel objects containing the soft segmentation(s), one per prediction class, \
+            List of target suffix associated with each prediction in `pred_list`
 
     """
+
+    # Define device
+    cuda_available, device = imed_utils.define_device(gpu_id)
 
     # Check if model folder exists and get filenames to be stored as string
     fname_model: str
@@ -363,76 +384,117 @@ def segment_volume(folder_model: str, fname_images: list, gpu_id: int = 0, optio
         set_postprocessing_options(options, context)
 
     # LOADER
-    loader_params = context["loader_parameters"]
-    slice_axis = imed_utils.AXIS_DCT[loader_params['slice_axis']]
+    loader_params = context[ConfigKW.LOADER_PARAMETERS]
+    slice_axis = imed_utils.AXIS_DCT[loader_params[LoaderParamsKW.SLICE_AXIS]]
     metadata = {}
     fname_roi = None
-    fname_prior = options['fname_prior'] if (options is not None) and ('fname_prior' in options) else None
+
+    if (options is not None) and (OptionKW.FNAME_PRIOR in options):
+        fname_prior = options.get(OptionKW.FNAME_PRIOR)
+    else:
+        fname_prior = None
+
     if fname_prior is not None:
-        if 'roi_params' in loader_params and loader_params['roi_params']['suffix'] is not None:
+        if LoaderParamsKW.ROI_PARAMS in loader_params and loader_params[LoaderParamsKW.ROI_PARAMS][ROIParamsKW.SUFFIX] is not None:
             fname_roi = fname_prior
         # TRANSFORMATIONS
         metadata = process_transformations(context, fname_roi, fname_prior, metadata, slice_axis, fname_images)
 
     # Compose transforms
-    _, _, transform_test_params = imed_transforms.get_subdatasets_transforms(context["transformation"])
+    _, _, transform_test_params = imed_transforms.get_subdatasets_transforms(context[ConfigKW.TRANSFORMATION])
 
     tranform_lst, undo_transforms = imed_transforms.prepare_transforms(transform_test_params)
 
     # Force filter_empty_mask to False if fname_roi = None
-    if fname_roi is None and 'filter_empty_mask' in loader_params["slice_filter_params"]:
+    if fname_roi is None and SliceFilterParamsKW.FILTER_EMPTY_MASK in loader_params[LoaderParamsKW.SLICE_FILTER_PARAMS]:
         logger.warning("fname_roi has not been specified, then the entire volume is processed.")
-        loader_params["slice_filter_params"]["filter_empty_mask"] = False
+        loader_params[LoaderParamsKW.SLICE_FILTER_PARAMS][SliceFilterParamsKW.FILTER_EMPTY_MASK] = False
 
-    kernel_3D = bool('Modified3DUNet' in context and context['Modified3DUNet']['applied']) or \
-                not context['default_model']['is_2d']
+    kernel_3D = bool(ConfigKW.MODIFIED_3D_UNET in context and context[ConfigKW.MODIFIED_3D_UNET][ModelParamsKW.APPLIED]) or \
+                not context[ConfigKW.DEFAULT_MODEL][ModelParamsKW.IS_2D]
+
+    if (options is not None) and (OptionKW.NO_PATCH in options) and kernel_3D:
+        logger.warning(f"The 'no-patch' option is provided but is not available for 3D models. "
+                       f"'no-patch' is ignored.")
+    if (options is not None) and (OptionKW.OVERLAP_2D in options) and kernel_3D:
+        logger.warning(f"The 'overlap-2d' option is provided but is not available for 3D models. "
+                       f"'overlap-2d' is ignored.")
 
     # Assign length_2D and stride_2D for 2D patching
-    length_2D = context["default_model"]["length_2D"] if "length_2D" in context["default_model"] else []
-    stride_2D = context["default_model"]["stride_2D"] if "stride_2D" in context["default_model"] else []
+    length_2D = context[ConfigKW.DEFAULT_MODEL][ModelParamsKW.LENGTH_2D] if \
+        ModelParamsKW.LENGTH_2D in context[ConfigKW.DEFAULT_MODEL] else []
+    stride_2D = context[ConfigKW.DEFAULT_MODEL][ModelParamsKW.STRIDE_2D] if \
+        ModelParamsKW.STRIDE_2D in context[ConfigKW.DEFAULT_MODEL] else []
+
     is_2d_patch = bool(length_2D)
+    if (options is not None) and (OptionKW.NO_PATCH in options) and not kernel_3D:
+        if is_2d_patch:
+            is_2d_patch = not options.get(OptionKW.NO_PATCH)
+            length_2D = []
+            stride_2D = []
+        else:
+            logger.warning(f"The 'no-patch' option is provided but the model has no 'length_2D' and "
+                           f"'stride_2D' parameters in its configuration file "
+                           f"'{fname_model_metadata.split('/')[-1]}'. 2D patching is ignored, the segmentation "
+                           f"is done on the entire image without patches.")
+        if OptionKW.OVERLAP_2D in options:
+            logger.warning(f"The 'no-patch' option is provided along with the 'overlap-2D' option. "
+                           f"2D patching is ignored, the segmentation is done on the entire image without patches.")
+    else:
+        if (options is not None) and (OptionKW.OVERLAP_2D in options) and not kernel_3D:
+            if (length_2D and stride_2D):
+                overlap_2D = options.get(OptionKW.OVERLAP_2D)
+                # Swap OverlapX and OverlapY resulting in an array in order [OverlapY, OverlapX]
+                # to match length_2D and stride_2D in [Height, Width] orientation.
+                overlap_2D[1], overlap_2D[0] = overlap_2D[0], overlap_2D[1]
+                # Adjust stride_2D with overlap_2D
+                stride_2D = [x1 - x2 for (x1, x2) in zip(length_2D, overlap_2D)]
+            else:
+                logger.warning(f"The 'overlap-2d' option is provided but the model has no 'length_2D' and "
+                               f"'stride_2D' parameters in its configuration file "
+                               f"'{fname_model_metadata.split('/')[-1]}'. 2D patching is ignored, the segmentation "
+                               f"is done on the entire image without patches.")
 
-    # Adjust stride_2D with overlap_2D option if present
-    if is_2d_patch and (options is not None) and ('overlap_2D' in options):
-        stride_2D = [x1 - x2 for (x1, x2) in zip(length_2D, options['overlap_2D'])]
-
-    # Add microscopy pixel size from options to metadata for filenames_pairs
-    if (options is not None) and ('pixel_size' in options):
-        metadata['PixelSize'] = options['pixel_size']
+    # Add microscopy pixel size and pixel size units from options to metadata for filenames_pairs
+    if (options is not None) and (OptionKW.PIXEL_SIZE in options):
+        metadata[MetadataKW.PIXEL_SIZE] = options.get(OptionKW.PIXEL_SIZE)
+    if (options is not None) and (OptionKW.PIXEL_SIZE_UNITS in options):
+        metadata[MetadataKW.PIXEL_SIZE_UNITS] = options.get(OptionKW.PIXEL_SIZE_UNITS)
 
     filename_pairs = [(fname_images, None, fname_roi, metadata if isinstance(metadata, list) else [metadata])]
 
     if kernel_3D:
         ds = MRI3DSubVolumeSegmentationDataset(filename_pairs,
                                                transform=tranform_lst,
-                                               length=context["Modified3DUNet"]["length_3D"],
-                                               stride=context["Modified3DUNet"]["stride_3D"])
-        logger.info(f"Loaded {len(ds)} {loader_params['slice_axis']} volumes of shape "
-                     f"{context['Modified3DUNet']['length_3D']}.")
+                                               length=context[ConfigKW.MODIFIED_3D_UNET][ModelParamsKW.LENGTH_3D],
+                                               stride=context[ConfigKW.MODIFIED_3D_UNET][ModelParamsKW.STRIDE_3D],
+                                               slice_axis=slice_axis)
+        logger.info(f"Loaded {len(ds)} {loader_params[LoaderParamsKW.SLICE_AXIS]} volumes of shape "
+                    f"{context[ConfigKW.MODIFIED_3D_UNET][ModelParamsKW.LENGTH_3D]}.")
     else:
         ds = MRI2DSegmentationDataset(filename_pairs,
                                       length=length_2D,
                                       stride=stride_2D,
                                       slice_axis=slice_axis,
-                                      cache=True,
+                                      nibabel_cache=True,
                                       transform=tranform_lst,
-                                      slice_filter_fn=imed_loader_utils.SliceFilter(
-                                          **loader_params["slice_filter_params"]))
+                                      slice_filter_fn=SliceFilter(
+                                          **loader_params[LoaderParamsKW.SLICE_FILTER_PARAMS]))
         ds.load_filenames()
         if is_2d_patch:
-            logger.info(f"Loaded {len(ds)} {loader_params['slice_axis']} patches of shape {length_2D}.")
+            logger.info(f"Loaded {len(ds)} {loader_params[LoaderParamsKW.SLICE_AXIS]} patches of shape {length_2D}.")
         else:
-            logger.info(f"Loaded {len(ds)} {loader_params['slice_axis']} slices.")
+            logger.info(f"Loaded {len(ds)} {loader_params[LoaderParamsKW.SLICE_AXIS]} slices.")
 
     model_params = {}
-    if 'FiLMedUnet' in context and context['FiLMedUnet']['applied']:
+    if ConfigKW.FILMED_UNET in context and context[ConfigKW.FILMED_UNET][ModelParamsKW.APPLIED]:
         onehotencoder = get_onehotencoder(context, folder_model, options, ds)
-        model_params.update({"name": 'FiLMedUnet',
-                             "film_onehotencoder": onehotencoder,
-                             "n_metadata": len([ll for l in onehotencoder.categories_ for ll in l])})
+        model_params.update({ModelParamsKW.NAME: ConfigKW.FILMED_UNET,
+                             ModelParamsKW.FILM_ONEHOTENCODER: onehotencoder,
+                             ModelParamsKW.N_METADATA: len([ll for l in onehotencoder.categories_ for ll in l])})
 
     # Data Loader
-    data_loader = DataLoader(ds, batch_size=context["training_parameters"]["batch_size"],
+    data_loader = DataLoader(ds, batch_size=context[ConfigKW.TRAINING_PARAMETERS][TrainingParamsKW.BATCH_SIZE],
                              shuffle=False, pin_memory=True,
                              collate_fn=imed_loader_utils.imed_collate,
                              num_workers=0)
@@ -441,10 +503,10 @@ def segment_volume(folder_model: str, fname_images: list, gpu_id: int = 0, optio
     preds_list, slice_idx_list = [], []
     last_sample_bool, weight_matrix, volume, image = False, None, None, None
     for i_batch, batch in enumerate(data_loader):
-        preds = get_preds(context, fname_model, model_params, gpu_id, batch)
+        preds = get_preds(context, fname_model, model_params, cuda_available, device, batch)
 
         # Set datatype to gt since prediction should be processed the same way as gt
-        for b in batch['input_metadata']:
+        for b in batch[MetadataKW.INPUT_METADATA]:
             for modality in b:
                 modality['data_type'] = 'gt'
 
@@ -469,7 +531,8 @@ def split_classes(nib_prediction):
     pred = nib_prediction.get_fdata()
     pred_list = []
     for c in range(pred.shape[-1]):
-        class_pred = nib.Nifti1Image(pred[..., c].astype('float32'), None, nib_prediction.header.copy())
+        class_pred = nib.Nifti1Image(pred[..., c].astype('float32'), nib_prediction.header.get_best_affine(),
+                                     nib_prediction.header.copy())
         pred_list.append(class_pred)
     return pred_list
 
@@ -478,24 +541,22 @@ def reconstruct_3d_object(context: dict, batch: dict, undo_transforms: UndoCompo
                           preds_list: list, kernel_3D: bool, is_2d_patch: bool, slice_axis: int, slice_idx_list: list,
                           data_loader: DataLoader, fname_images: list, i_batch: int, last_sample_bool: bool,
                           weight_matrix: tensor, volume: tensor, image: tensor):
-    """Reconstructs the 3D object from the current batch, and returns the list of predictions and
-       targets.
+    """Reconstructs the 3D object from the current batch, and returns the list of predictions and targets.
 
     Args:
-
         context (dict): configuration dict.
         batch (dict): Dictionary containing input, gt and metadata
         undo_transforms (UndoCompose): Undo transforms so prediction match original image resolution and shape
         preds (tensor): Subvolume predictions
         preds_list (list of tensor): list of subvolume predictions.
         kernel_3D (bool): true when using 3D kernel.
-        is_2d_patch (bool): True if length in default model params.
+        is_2d_patch (bool): Indicates if 2d patching is used.
         slice_axis (int): Indicates the axis used for the 2D slice extraction: Sagittal: 0, Coronal: 1, Axial: 2.
         slice_idx_list (list of int): list of indices for the axis slices.
         data_loader (DataLoader): DataLoader object containing batches using in object construction.
         fname_images (list): list of image filenames (e.g. .nii.gz) to segment.
         i_batch (int): index of current batch.
-        last_sample_bool: : flag to indicate whether this is the last sample in the 3D volume
+        last_sample_bool (bool) : flag to indicate whether this is the last sample in the 3D volume
         weight_matrix (tensor): the weight matrix
         volume (tensor): the volume tensor that is being partially reconstructed through the loop
         image (tensor): the image tensor that is being partially reconstructed through the loop
@@ -505,22 +566,23 @@ def reconstruct_3d_object(context: dict, batch: dict, undo_transforms: UndoCompo
         target_list (list): list of targets
         last_sample_bool (bool): flag to indicate whether this is the last sample in the 3D volume
         weight_matrix (tensor): the weight matrix. Must be returned as passing tensor by reference is NOT reliable.
-        volume (tensor): the volume tensor that is being partially reconstructed through the loop. Must be returned
-         as passing tensor by reference is NOT reliable.
-        image (tensor): the vimage tensor that is being partially reconstructed through the loop. Must be returned
-         as passing tensor by reference is NOT reliable.
+        volume (tensor): the volume tensor that is being partially reconstructed through the loop. Must be returned \
+            as passing tensor by reference is NOT reliable.
+        image (tensor): the vimage tensor that is being partially reconstructed through the loop. Must be returned \
+            as passing tensor by reference is NOT reliable.
     """
     pred_list = []
     target_list = []
     for i_slice in range(len(preds)):
-        if "bounding_box" in batch['input_metadata'][i_slice][0]:
+        if "bounding_box" in batch[MetadataKW.INPUT_METADATA][i_slice][0]:
             imed_obj_detect.adjust_undo_transforms(undo_transforms.transforms, batch, i_slice)
 
-        batch['gt_metadata'] = [[metadata[0]] * preds.shape[1] for metadata in batch['input_metadata']]
+        batch[MetadataKW.GT_METADATA] = [[metadata[0]] * preds.shape[1] for metadata in batch[MetadataKW.INPUT_METADATA]]
         if kernel_3D:
             preds_undo, metadata, last_sample_bool, volume, weight_matrix = \
                 volume_reconstruction(batch, preds, undo_transforms, i_slice, volume, weight_matrix)
-            preds_list = [np.array(preds_undo)]
+            if last_sample_bool:
+                preds_list = [np.array(preds_undo)]
         else:
             if is_2d_patch:
                 # undo transformations for patch and reconstruct slice
@@ -531,16 +593,16 @@ def reconstruct_3d_object(context: dict, batch: dict, undo_transforms: UndoCompo
                     # Add new segmented slice to preds_list
                     preds_list.append(np.array(preds_i_undo))
                     # Store the slice index of preds_i_undo in the original 3D image
-                    slice_idx_list.append(int(batch['input_metadata'][i_slice][0]['slice_index']))
+                    slice_idx_list.append(int(batch[MetadataKW.INPUT_METADATA][i_slice][0]['slice_index']))
             else:
                 # undo transformations for slice
                 preds_i_undo, metadata_idx = undo_transforms(preds[i_slice],
-                                                             batch["gt_metadata"][i_slice],
+                                                             batch[MetadataKW.GT_METADATA][i_slice],
                                                              data_type='gt')
                 # Add new segmented slice to preds_list
                 preds_list.append(np.array(preds_i_undo))
                 # Store the slice index of preds_i_undo in the original 3D image
-                slice_idx_list.append(int(batch['input_metadata'][i_slice][0]['slice_index']))
+                slice_idx_list.append(int(batch[MetadataKW.INPUT_METADATA][i_slice][0]['slice_index']))
 
         # If last batch and last sample of this batch, then reconstruct 3D object
         if (i_batch == len(data_loader) - 1 and i_slice == len(batch['gt']) - 1) or last_sample_bool:
@@ -552,10 +614,10 @@ def reconstruct_3d_object(context: dict, batch: dict, undo_transforms: UndoCompo
                                    kernel_dim='3d' if kernel_3D else '2d',
                                    debug=False,
                                    bin_thr=-1,
-                                   postprocessing=context['postprocessing'])
+                                   postprocessing=context[ConfigKW.POSTPROCESSING])
 
             pred_list = split_classes(pred_nib)
-            target_list = context['loader_parameters']['target_suffix']
+            target_list = context[ConfigKW.LOADER_PARAMETERS][LoaderParamsKW.TARGET_SUFFIX]
 
     return pred_list, target_list, last_sample_bool, weight_matrix, volume, image
 
@@ -579,7 +641,8 @@ def volume_reconstruction(batch: dict, pred: tensor, undo_transforms: UndoCompos
         volume (tensor): representing the volume reconstructed
         weight_matrix (tensor): weight matrix
     """
-    x_min, x_max, y_min, y_max, z_min, z_max = batch['input_metadata'][smp_idx][0]['coord']
+    pred_undo, metadata = None, None
+    x_min, x_max, y_min, y_max, z_min, z_max = batch[MetadataKW.INPUT_METADATA][smp_idx][0]['coord']
     num_pred = pred[smp_idx].shape[0]
 
     # A boolean flag indicate whether the current volume is the VERY first subvolume of the entire 3D volume/space.
@@ -587,7 +650,7 @@ def volume_reconstruction(batch: dict, pred: tensor, undo_transforms: UndoCompos
     first_sample: bool = (x_min == 0 and y_min == 0 and z_min == 0)
 
     # Get the Dimension
-    x, y, z = batch['input_metadata'][smp_idx][0]['index_shape']
+    x, y, z = batch[MetadataKW.INPUT_METADATA][smp_idx][0]['index_shape']
 
     # If this is the first sample, instantiate a ZERO tensor based on the dimension
     if first_sample:
@@ -602,10 +665,9 @@ def volume_reconstruction(batch: dict, pred: tensor, undo_transforms: UndoCompos
 
     if last_sample_bool:
         volume /= weight_matrix
-
-    pred_undo, metadata = undo_transforms(volume,
-                                          batch['gt_metadata'][smp_idx],
-                                          data_type='gt')
+        pred_undo, metadata = undo_transforms(volume,
+                                              batch[MetadataKW.GT_METADATA][smp_idx],
+                                              data_type='gt')
     return pred_undo, metadata, last_sample_bool, volume, weight_matrix
 
 
@@ -622,13 +684,14 @@ def image_reconstruction(batch: dict, pred: tensor, undo_transforms: UndoCompose
         weight_matrix (tensor): Weights containing the number of predictions for each pixel
 
     Returns:
-        pred_undo (tensor): undone patch,
-        metadata (dict): metadata,
-        last_sample_bool (bool): boolean representing if its the last patch of the image
+        pred_undo (tensor): undone image
+        metadata (dict): metadata
+        last_patch_bool (bool): boolean representing if its the last patch of the image
         image (tensor): representing the image reconstructed
         weight_matrix (tensor): weight matrix
     """
-    x_min, x_max, y_min, y_max = batch['input_metadata'][smp_idx][0]['coord']
+    pred_undo, metadata = None, None
+    x_min, x_max, y_min, y_max = batch[MetadataKW.INPUT_METADATA][smp_idx][0]['coord']
     num_pred = pred[smp_idx].shape[0]
 
     # A boolean flag indicate whether the current patch is the VERY first patch of the entire 2D image.
@@ -636,7 +699,7 @@ def image_reconstruction(batch: dict, pred: tensor, undo_transforms: UndoCompose
     first_patch: bool = (x_min == 0 and y_min == 0)
 
     # Get the Dimension
-    x, y = batch['input_metadata'][smp_idx][0]['index_shape']
+    x, y = batch[MetadataKW.INPUT_METADATA][smp_idx][0]['index_shape']
 
     # If this is the first sample, instantiate a ZERO tensor based on the dimension
     if first_patch:
@@ -650,6 +713,6 @@ def image_reconstruction(batch: dict, pred: tensor, undo_transforms: UndoCompose
     weight_matrix[:, x_min:x_max, y_min:y_max] += 1
     if last_patch_bool:
         image /= weight_matrix
+        pred_undo, metadata = undo_transforms(image, batch[MetadataKW.GT_METADATA][smp_idx], data_type='gt')
 
-    pred_undo, metadata = undo_transforms(image, batch['gt_metadata'][smp_idx], data_type='gt')
     return pred_undo, metadata, last_patch_bool, image, weight_matrix
