@@ -1,5 +1,4 @@
 import json
-import os
 import argparse
 import copy
 import joblib
@@ -10,6 +9,7 @@ import platform
 import multiprocessing
 import re
 
+from ivadomed.loader.bids_dataframe import BidsDataframe
 from ivadomed import evaluation as imed_evaluation
 from ivadomed import config_manager as imed_config_manager
 from ivadomed import testing as imed_testing
@@ -19,7 +19,11 @@ from ivadomed import utils as imed_utils
 from ivadomed import metrics as imed_metrics
 from ivadomed import inference as imed_inference
 from ivadomed.loader import utils as imed_loader_utils, loader as imed_loader, film as imed_film
+from ivadomed.keywords import ConfigKW, ModelParamsKW, LoaderParamsKW, ContrastParamsKW, BalanceSamplesKW, \
+    TrainingParamsKW, ObjectDetectionParamsKW, UncertaintyKW, PostprocessingKW, BinarizeProdictionKW, MetricsKW, \
+    MetadataKW, OptionKW, SplitDatasetKW
 from loguru import logger
+from pathlib import Path
 
 cudnn.benchmark = True
 
@@ -65,6 +69,14 @@ def get_parser():
     optional_args.add_argument('--resume-training', dest="resume_training", required=False, action='store_true',
                                help='Load a saved model ("checkpoint.pth.tar" in the output directory specified either with flag "--path-output" or via the config file "output_path" argument)  '
                                     'for resume training. This training state is saved everytime a new best model is saved in the output directory specified with flag "--path-output"')
+    optional_args.add_argument('--no-patch', dest="no_patch", action='store_true', required=False,
+                               help='2D patches are not used while segmenting with models trained with patches '
+                               '(command "--segment" only). The "--no-patch" argument supersedes the "--overlap-2D" argument. '
+                               ' This option may not be suitable with large images depending on computer RAM capacity.')
+    optional_args.add_argument('--overlap-2d', dest="overlap_2d", required=False, type=int, nargs="+",
+                                help='Custom overlap for 2D patches while segmenting (command "--segment" only). '
+                                'Example: "--overlap-2d 48 48" for an overlap of 48 pixels between patches in X and Y respectively. '
+                                'Default model overlap is used otherwise.')
     optional_args.add_argument('-h', '--help', action='help', default=argparse.SUPPRESS,
                                help='Shows function documentation.')
 
@@ -72,43 +84,43 @@ def get_parser():
 
 
 def create_path_model(context, model_params, ds_train, path_output, train_onehotencoder):
-    path_model = os.path.join(path_output, context["model_name"])
-    if not os.path.isdir(path_model):
-        logger.info('Creating model directory: {}'.format(path_model))
-        os.makedirs(path_model)
-        if 'film_layers' in model_params and any(model_params['film_layers']):
-            joblib.dump(train_onehotencoder, os.path.join(path_model, "one_hot_encoder.joblib"))
-            if 'metadata_dict' in ds_train[0]['input_metadata'][0]:
-                metadata_dict = ds_train[0]['input_metadata'][0]['metadata_dict']
-                joblib.dump(metadata_dict, os.path.join(path_model, "metadata_dict.joblib"))
+    path_model = Path(path_output, context[ConfigKW.MODEL_NAME])
+    if not path_model.is_dir():
+        logger.info(f'Creating model directory: {path_model}')
+        path_model.mkdir(parents=True)
+        if ModelParamsKW.FILM_LAYERS in model_params and any(model_params[ModelParamsKW.FILM_LAYERS]):
+            joblib.dump(train_onehotencoder, path_model.joinpath("one_hot_encoder.joblib"))
+            if MetadataKW.METADATA_DICT in ds_train[0][MetadataKW.INPUT_METADATA][0]:
+                metadata_dict = ds_train[0][MetadataKW.INPUT_METADATA][0][MetadataKW.METADATA_DICT]
+                joblib.dump(metadata_dict, path_model.joinpath("metadata_dict.joblib"))
 
     else:
-        logger.info('Model directory already exists: {}'.format(path_model))
+        logger.info(f'Model directory already exists: {path_model}')
 
 
 def check_multiple_raters(is_train, loader_params):
-    if any([isinstance(class_suffix, list) for class_suffix in loader_params["target_suffix"]]):
+    if any([isinstance(class_suffix, list) for class_suffix in loader_params[LoaderParamsKW.TARGET_SUFFIX]]):
         logger.info(
             "Annotations from multiple raters will be used during model training, one annotation from one rater "
             "randomly selected at each iteration.\n")
         if not is_train:
             logger.error(
                 "Please provide only one annotation per class in 'target_suffix' when not training a model.\n")
-            exit()
+            sys.exit()
 
 
 def film_normalize_data(context, model_params, ds_train, ds_valid, path_output):
     # Normalize metadata before sending to the FiLM network
     results = imed_film.get_film_metadata_models(ds_train=ds_train,
-                                                 metadata_type=model_params['metadata'],
-                                                 debugging=context["debugging"])
+                                                 metadata_type=model_params[ModelParamsKW.METADATA],
+                                                 debugging=context[ConfigKW.DEBUGGING])
     ds_train, train_onehotencoder, metadata_clustering_models = results
-    ds_valid = imed_film.normalize_metadata(ds_valid, metadata_clustering_models, context["debugging"],
-                                            model_params['metadata'])
-    model_params.update({"film_onehotencoder": train_onehotencoder,
-                         "n_metadata": len([ll for l in train_onehotencoder.categories_ for ll in l])})
-    joblib.dump(metadata_clustering_models, os.path.join(path_output, "clustering_models.joblib"))
-    joblib.dump(train_onehotencoder, os.path.join(path_output + "one_hot_encoder.joblib"))
+    ds_valid = imed_film.normalize_metadata(ds_valid, metadata_clustering_models, context[ConfigKW.DEBUGGING],
+                                            model_params[ModelParamsKW.METADATA])
+    model_params.update({ModelParamsKW.FILM_ONEHOTENCODER: train_onehotencoder,
+                         ModelParamsKW.N_METADATA: len([ll for l in train_onehotencoder.categories_ for ll in l])})
+    joblib.dump(metadata_clustering_models, Path(path_output, "clustering_models.joblib"))
+    joblib.dump(train_onehotencoder, Path(path_output + "one_hot_encoder.joblib"))
 
     return model_params, ds_train, ds_valid, train_onehotencoder
 
@@ -124,109 +136,121 @@ def get_dataset(bids_df, loader_params, data_lst, transform_params, cuda_availab
 def save_config_file(context, path_output):
     # Save config file within path_output and path_output/model_name
     # Done after the threshold_analysis to propate this info in the config files
-    with open(os.path.join(path_output, "config_file.json"), 'w') as fp:
+    with Path(path_output, "config_file.json").open(mode='w') as fp:
         json.dump(context, fp, indent=4)
-    with open(os.path.join(path_output, context["model_name"], context["model_name"] + ".json"), 'w') as fp:
+    with Path(path_output, context[ConfigKW.MODEL_NAME], context[ConfigKW.MODEL_NAME] + ".json").open(mode='w') as fp:
         json.dump(context, fp, indent=4)
 
 
 def set_loader_params(context, is_train):
-    loader_params = copy.deepcopy(context["loader_parameters"])
+    loader_params = copy.deepcopy(context[ConfigKW.LOADER_PARAMETERS])
     if is_train:
-        loader_params["contrast_params"]["contrast_lst"] = loader_params["contrast_params"]["training_validation"]
+        loader_params[LoaderParamsKW.CONTRAST_PARAMS][ContrastParamsKW.CONTRAST_LST] = \
+            loader_params[LoaderParamsKW.CONTRAST_PARAMS][ContrastParamsKW.TRAINING_VALIDATION]
     else:
-        loader_params["contrast_params"]["contrast_lst"] = loader_params["contrast_params"]["testing"]
-    if "FiLMedUnet" in context and context["FiLMedUnet"]["applied"]:
-        loader_params.update({"metadata_type": context["FiLMedUnet"]["metadata"]})
+        loader_params[LoaderParamsKW.CONTRAST_PARAMS][ContrastParamsKW.CONTRAST_LST] =\
+            loader_params[LoaderParamsKW.CONTRAST_PARAMS][ContrastParamsKW.TESTING]
+    if ConfigKW.FILMED_UNET in context and context[ConfigKW.FILMED_UNET][ModelParamsKW.APPLIED]:
+        loader_params.update({LoaderParamsKW.METADATA_TYPE: context[ConfigKW.FILMED_UNET][ModelParamsKW.METADATA]})
 
     # Load metadata necessary to balance the loader
-    if context['training_parameters']['balance_samples']['applied'] and \
-            context['training_parameters']['balance_samples']['type'] != 'gt':
-        loader_params.update({"metadata_type": context['training_parameters']['balance_samples']['type']})
+    if context[ConfigKW.TRAINING_PARAMETERS][TrainingParamsKW.BALANCE_SAMPLES][BalanceSamplesKW.APPLIED] and \
+            context[ConfigKW.TRAINING_PARAMETERS][TrainingParamsKW.BALANCE_SAMPLES][BalanceSamplesKW.TYPE] != 'gt':
+        loader_params.update({LoaderParamsKW.METADATA_TYPE:
+                                  context[ConfigKW.TRAINING_PARAMETERS][TrainingParamsKW.BALANCE_SAMPLES][BalanceSamplesKW.TYPE]})
     return loader_params
 
 
 def set_model_params(context, loader_params):
-    model_params = copy.deepcopy(context["default_model"])
-    model_params["folder_name"] = copy.deepcopy(context["model_name"])
+    model_params = copy.deepcopy(context[ConfigKW.DEFAULT_MODEL])
+    model_params[ModelParamsKW.FOLDER_NAME] = copy.deepcopy(context[ConfigKW.MODEL_NAME])
     model_context_list = [model_name for model_name in MODEL_LIST
-                          if model_name in context and context[model_name]["applied"]]
+                          if model_name in context and context[model_name][ModelParamsKW.APPLIED]]
     if len(model_context_list) == 1:
-        model_params["name"] = model_context_list[0]
+        model_params[ModelParamsKW.NAME] = model_context_list[0]
         model_params.update(context[model_context_list[0]])
-    elif 'Modified3DUNet' in model_context_list and 'FiLMedUnet' in model_context_list and len(model_context_list) == 2:
-        model_params["name"] = 'Modified3DUNet'
+    elif ConfigKW.MODIFIED_3D_UNET in model_context_list and ConfigKW.FILMED_UNET in model_context_list \
+            and len(model_context_list) == 2:
+        model_params[ModelParamsKW.NAME] = ConfigKW.MODIFIED_3D_UNET
         for i in range(len(model_context_list)):
             model_params.update(context[model_context_list[i]])
     elif len(model_context_list) > 1:
-        logger.error('ERROR: Several models are selected in the configuration file: {}.'
-              'Please select only one (i.e. only one where: "applied": true).'.format(model_context_list))
+        logger.error(f'ERROR: Several models are selected in the configuration file: {model_context_list}.'
+              'Please select only one (i.e. only one where: "applied": true).')
         exit()
 
-    model_params['is_2d'] = False if "Modified3DUNet" in model_params['name'] else model_params['is_2d']
+    model_params[ModelParamsKW.IS_2D] = False if ConfigKW.MODIFIED_3D_UNET in model_params[ModelParamsKW.NAME] \
+        else model_params[ModelParamsKW.IS_2D]
     # Get in_channel from contrast_lst
-    if loader_params["multichannel"]:
-        model_params["in_channel"] = len(loader_params["contrast_params"]["contrast_lst"])
+    if loader_params[LoaderParamsKW.MULTICHANNEL]:
+        model_params[ModelParamsKW.IN_CHANNEL] = \
+            len(loader_params[LoaderParamsKW.CONTRAST_PARAMS][ContrastParamsKW.CONTRAST_LST])
     else:
-        model_params["in_channel"] = 1
+        model_params[ModelParamsKW.IN_CHANNEL] = 1
     # Get out_channel from target_suffix
-    model_params["out_channel"] = len(loader_params["target_suffix"])
+    model_params[ModelParamsKW.OUT_CHANNEL] = len(loader_params[LoaderParamsKW.TARGET_SUFFIX])
     # If multi-class output, then add background class
-    if model_params["out_channel"] > 1:
-        model_params.update({"out_channel": model_params["out_channel"] + 1})
+    if model_params[ModelParamsKW.OUT_CHANNEL] > 1:
+        model_params.update({ModelParamsKW.OUT_CHANNEL: model_params[ModelParamsKW.OUT_CHANNEL] + 1})
     # Display for spec' check
     imed_utils.display_selected_model_spec(params=model_params)
     # Update loader params
-    if 'object_detection_params' in context:
-        object_detection_params = context['object_detection_params']
-        object_detection_params.update({"gpu_ids": context['gpu_ids'][0],
-                                        "path_output": context['path_output']})
-        loader_params.update({"object_detection_params": object_detection_params})
+    if ConfigKW.OBJECT_DETECTION_PARAMS in context:
+        object_detection_params = context[ConfigKW.OBJECT_DETECTION_PARAMS]
+        object_detection_params.update({ObjectDetectionParamsKW.GPU_IDS: context[ConfigKW.GPU_IDS][0],
+                                        ObjectDetectionParamsKW.PATH_OUTPUT: context[ConfigKW.PATH_OUTPUT]})
+        loader_params.update({ConfigKW.OBJECT_DETECTION_PARAMS: object_detection_params})
 
-    loader_params.update({"model_params": model_params})
+    loader_params.update({LoaderParamsKW.MODEL_PARAMS: model_params})
 
     return model_params, loader_params
 
 
 def set_output_path(context):
-    path_output = copy.deepcopy(context["path_output"])
-    if not os.path.isdir(path_output):
-        logger.info('Creating output path: {}'.format(path_output))
-        os.makedirs(path_output)
+    path_output = copy.deepcopy(context[ConfigKW.PATH_OUTPUT])
+    if not Path(path_output).is_dir():
+        logger.info(f'Creating output path: {path_output}')
+        Path(path_output).mkdir(parents=True)
     else:
-        logger.info('Output path already exists: {}'.format(path_output))
+        logger.info(f'Output path already exists: {path_output}')
 
     return path_output
 
 
 def update_film_model_params(context, ds_test, model_params, path_output):
-    clustering_path = os.path.join(path_output, "clustering_models.joblib")
+    clustering_path = Path(path_output, "clustering_models.joblib")
     metadata_clustering_models = joblib.load(clustering_path)
     # Model directory
-    ohe_path = os.path.join(path_output, context["model_name"], "one_hot_encoder.joblib")
+    ohe_path = Path(path_output, context[ConfigKW.MODEL_NAME], "one_hot_encoder.joblib")
     one_hot_encoder = joblib.load(ohe_path)
-    ds_test = imed_film.normalize_metadata(ds_test, metadata_clustering_models, context["debugging"],
-                                           model_params['metadata'])
-    model_params.update({"film_onehotencoder": one_hot_encoder,
-                         "n_metadata": len([ll for l in one_hot_encoder.categories_ for ll in l])})
+    ds_test = imed_film.normalize_metadata(ds_test, metadata_clustering_models, context[ConfigKW.DEBUGGING],
+                                           model_params[ModelParamsKW.METADATA])
+    model_params.update({ModelParamsKW.FILM_ONEHOTENCODER: one_hot_encoder,
+                         ModelParamsKW.N_METADATA: len([ll for l in one_hot_encoder.categories_ for ll in l])})
 
     return ds_test, model_params
 
 
-def run_segment_command(context, model_params):
+def run_segment_command(context, model_params, no_patch, overlap_2d):
     # BIDSDataframe of all image files
     # Indexing of derivatives is False for command segment
-    bids_df = imed_loader_utils.BidsDataframe(context['loader_parameters'], context['path_output'], derivatives=False)
+    # split_method is unused for command segment
+    bids_df = BidsDataframe(
+        context.get(ConfigKW.LOADER_PARAMETERS),
+        context.get(ConfigKW.PATH_OUTPUT),
+        derivatives=False,
+        split_method=None
+    )
 
     # Append subjects filenames into a list
-    bids_subjects = sorted(bids_df.df['filename'].to_list())
+    bids_subjects = sorted(bids_df.df.get('filename').to_list())
 
     # Add postprocessing to packaged model
-    path_model = os.path.join(context['path_output'], context['model_name'])
-    path_model_config = os.path.join(path_model, context['model_name'] + ".json")
-    model_config = imed_config_manager.load_json(path_model_config)
-    model_config['postprocessing'] = context['postprocessing']
-    with open(path_model_config, 'w') as fp:
+    path_model = Path(context[ConfigKW.PATH_OUTPUT], context[ConfigKW.MODEL_NAME])
+    path_model_config = Path(path_model, context[ConfigKW.MODEL_NAME] + ".json")
+    model_config = imed_config_manager.load_json(str(path_model_config))
+    model_config[ConfigKW.POSTPROCESSING] = context.get(ConfigKW.POSTPROCESSING)
+    with path_model_config.open(mode='w') as fp:
         json.dump(model_config, fp, indent=4)
     options = {}
 
@@ -234,7 +258,7 @@ def run_segment_command(context, model_params):
     seen_subj_ids = []
 
     for subject in bids_subjects:
-        if context['loader_parameters']['multichannel']:
+        if context.get(ConfigKW.LOADER_PARAMETERS).get(LoaderParamsKW.MULTICHANNEL):
             # Get subject_id for multichannel
             df_sub = bids_df.df.loc[bids_df.df['filename'] == subject]
             subj_id = re.sub(r'_' + df_sub['suffix'].values[0] + '.*', '', subject)
@@ -242,7 +266,7 @@ def run_segment_command(context, model_params):
                 # if subj_id has not been seen yet
                 fname_img = []
                 provided_contrasts = []
-                contrasts = context['loader_parameters']['contrast_params']['testing']
+                contrasts = context[ConfigKW.LOADER_PARAMETERS][LoaderParamsKW.CONTRAST_PARAMS][ContrastParamsKW.TESTING]
                 # Keep contrast order
                 for c in contrasts:
                     df_tmp = bids_df.df[
@@ -252,8 +276,8 @@ def run_segment_command(context, model_params):
                         fname_img.append(df_tmp['path'].values[0])
                 seen_subj_ids.append(subj_id)
                 if len(fname_img) != len(contrasts):
-                    logger.warning("Missing contrast for subject {}. {} were provided but {} are required. Skipping "
-                                   "subject.".format(subj_id, provided_contrasts, contrasts))
+                    logger.warning(f"Missing contrast for subject {subj_id}. {provided_contrasts} were provided but "
+                                   f"{contrasts} are required. Skipping subject.")
                     continue
             else:
                 # Returns an empty list for subj_id already seen
@@ -262,36 +286,51 @@ def run_segment_command(context, model_params):
             fname_img = bids_df.df[bids_df.df['filename'] == subject]['path'].to_list()
 
         # Add film metadata to options for segment_volume
-        if 'film_layers' in model_params and any(model_params['film_layers']) and model_params['metadata']:
-            metadata = bids_df.df[bids_df.df['filename'] == subject][model_params['metadata']].values[0]
-            options['metadata'] = metadata
+        if ModelParamsKW.FILM_LAYERS in model_params and any(model_params[ModelParamsKW.FILM_LAYERS]) \
+                and model_params[ModelParamsKW.METADATA]:
+            metadata = bids_df.df[bids_df.df['filename'] == subject][model_params[ModelParamsKW.METADATA]].values[0]
+            options[OptionKW.METADATA] = metadata
 
-        # Add microscopy pixel size metadata to options for segment_volume
-        if 'PixelSize' in bids_df.df.columns:
-            options['pixel_size'] = bids_df.df.loc[bids_df.df['filename'] == subject]['PixelSize'].values[0]
+        # Add microscopy pixel size and pixel size units metadata to options for segment_volume
+        if MetadataKW.PIXEL_SIZE in bids_df.df.columns:
+            options[OptionKW.PIXEL_SIZE] = \
+                bids_df.df.loc[bids_df.df['filename'] == subject][MetadataKW.PIXEL_SIZE].values[0]
+        if MetadataKW.PIXEL_SIZE_UNITS in bids_df.df.columns:
+            options[OptionKW.PIXEL_SIZE_UNITS] = \
+                bids_df.df.loc[bids_df.df['filename'] == subject][MetadataKW.PIXEL_SIZE_UNITS].values[0]
+
+        # Add 'no_patch' and 'overlap-2d' argument to options
+        if no_patch:
+            options[OptionKW.NO_PATCH] = no_patch
+        if overlap_2d:
+            options[OptionKW.OVERLAP_2D] = overlap_2d
 
         if fname_img:
-            pred_list, target_list = imed_inference.segment_volume(path_model,
+            pred_list, target_list = imed_inference.segment_volume(str(path_model),
                                                                    fname_images=fname_img,
-                                                                   gpu_id=context['gpu_ids'][0],
+                                                                   gpu_id=context[ConfigKW.GPU_IDS][0],
                                                                    options=options)
-            pred_path = os.path.join(context['path_output'], "pred_masks")
-            if not os.path.exists(pred_path):
-                os.makedirs(pred_path)
+            pred_path = Path(context[ConfigKW.PATH_OUTPUT], "pred_masks")
+            if not pred_path.exists():
+                pred_path.mkdir(parents=True)
+
+            # Reformat target list to include class index and be compatible with multiple raters
+            target_list = ["_class-%d" % i for i in range(len(target_list))]
 
             for pred, target in zip(pred_list, target_list):
                 filename = subject.split('.')[0] + target + "_pred" + ".nii.gz"
-                nib.save(pred, os.path.join(pred_path, filename))
+                nib.save(pred, Path(pred_path, filename))
 
             # For Microscopy PNG/TIF files (TODO: implement OMETIFF behavior)
             extension = imed_loader_utils.get_file_extension(subject)
             if "nii" not in extension:
                 imed_inference.pred_to_png(pred_list,
                                            target_list,
-                                           os.path.join(pred_path, subject).replace(extension, ''))
+                                           str(Path(pred_path, subject)).replace(extension, ''),
+                                           suffix="_pred.png")
 
 
-def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
+def run_command(context, n_gif=0, thr_increment=None, resume_training=False, no_patch=False, overlap_2d=None):
     """Run main command.
 
     This function is central in the ivadomed project as training / testing / evaluation commands
@@ -306,8 +345,14 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
         thr_increment (float): A threshold analysis is performed at the end of the training using the trained model and
             the training + validation sub-dataset to find the optimal binarization threshold. The specified value
             indicates the increment between 0 and 1 used during the ROC analysis (e.g. 0.1).
-        resume_training (bool): Load a saved model ("checkpoint.pth.tar" in the output directory specified with flag "--path-output" or via the config file "output_path" '            This training state is saved everytime a new best model is saved in the log
-            argument) for resume training directory.
+        resume_training (bool): Load a saved model ("checkpoint.pth.tar" in the output directory specified with flag
+            "--path-output" or via the config file "output_path". This training state is saved everytime a new best
+            model is saved in the log argument) for resume training directory.
+        no_patch (bool): If True, 2D patches are not used while segmenting with models trained with patches
+            (command "--segment" only). Default: False (i.e. segment with patches). The "no_patch" option supersedes
+            the "overlap_2D" option.
+        overlap_2d (list of int): Custom overlap for 2D patches while segmenting (command "--segment" only).
+            Default model overlap is used otherwise.
 
     Returns:
         float or pandas.DataFrame or None:
@@ -317,60 +362,65 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
             * If "segment" command: No return value.
 
     """
-    command = copy.deepcopy(context["command"])
+    command = copy.deepcopy(context[ConfigKW.COMMAND])
     path_output = set_output_path(context)
-    log_file = os.path.join(context['path_output'], context['log_file'])
+    path_log = Path(context.get('path_output'), context.get('log_file'))
     logger.remove()
-    logger.add(log_file)
+    logger.add(str(path_log))
     logger.add(sys.stdout)
 
     # Create a log with the version of the Ivadomed software and the version of the Annexed dataset (if present)
     create_dataset_and_ivadomed_version_log(context)
 
-    cuda_available, device = imed_utils.define_device(context['gpu_ids'][0])
+    cuda_available, device = imed_utils.define_device(context[ConfigKW.GPU_IDS][0])
 
     # BACKWARDS COMPATIBILITY: If bids_path is string, assign to list - Do this here so it propagates to all functions
-    context['loader_parameters']['path_data'] = imed_utils.format_path_data(context['loader_parameters']['path_data'])
+    context[ConfigKW.LOADER_PARAMETERS][LoaderParamsKW.PATH_DATA] =\
+        imed_utils.format_path_data(context[ConfigKW.LOADER_PARAMETERS][LoaderParamsKW.PATH_DATA])
 
     # Loader params
     loader_params = set_loader_params(context, command == "train")
 
     # Get transforms for each subdataset
     transform_train_params, transform_valid_params, transform_test_params = \
-        imed_transforms.get_subdatasets_transforms(context["transformation"])
+        imed_transforms.get_subdatasets_transforms(context[ConfigKW.TRANSFORMATION])
 
     # MODEL PARAMETERS
     model_params, loader_params = set_model_params(context, loader_params)
 
     if command == 'segment':
-        run_segment_command(context, model_params)
+        run_segment_command(context, model_params, no_patch, overlap_2d)
         return
 
     # BIDSDataframe of all image files
-    # Indexing of derivatives is True for command train and test
-    bids_df = imed_loader_utils.BidsDataframe(loader_params, path_output, derivatives=True)
+    # Indexing of derivatives is True for commands train and test
+    # split_method is used for removing unused subject files in bids_df for commands train and test
+    bids_df = BidsDataframe(loader_params, path_output, derivatives=True,
+        split_method=context.get(ConfigKW.SPLIT_DATASET).get(SplitDatasetKW.SPLIT_METHOD))
 
     # Get subject filenames lists. "segment" command uses all participants of data path, hence no need to split
-    train_lst, valid_lst, test_lst = imed_loader_utils.get_subdatasets_subject_files_list(context["split_dataset"],
+    train_lst, valid_lst, test_lst = imed_loader_utils.get_subdatasets_subject_files_list(context[ConfigKW.SPLIT_DATASET],
                                                                                           bids_df.df,
                                                                                           path_output,
-                                                                                          context["loader_parameters"]
-                                                                                          ['subject_selection'])
+                                                                                          context.get(ConfigKW.LOADER_PARAMETERS).get(
+                                                                                              LoaderParamsKW.SUBJECT_SELECTION))
 
     # Generating sha256 for the training files
     imed_utils.generate_sha_256(context, bids_df.df, train_lst)
 
     # TESTING PARAMS
     # Aleatoric uncertainty
-    if context['uncertainty']['aleatoric'] and context['uncertainty']['n_it'] > 0:
+    if context[ConfigKW.UNCERTAINTY][UncertaintyKW.ALEATORIC] \
+            and context[ConfigKW.UNCERTAINTY][UncertaintyKW.N_IT] > 0:
         transformation_dict = transform_train_params
     else:
         transformation_dict = transform_test_params
     undo_transforms = imed_transforms.UndoCompose(imed_transforms.Compose(transformation_dict, requires_undo=True))
-    testing_params = copy.deepcopy(context["training_parameters"])
-    testing_params.update({'uncertainty': context["uncertainty"]})
-    testing_params.update({'target_suffix': loader_params["target_suffix"], 'undo_transforms': undo_transforms,
-                           'slice_axis': loader_params['slice_axis']})
+    testing_params = copy.deepcopy(context[ConfigKW.TRAINING_PARAMETERS])
+    testing_params.update({ConfigKW.UNCERTAINTY: context[ConfigKW.UNCERTAINTY]})
+    testing_params.update({LoaderParamsKW.TARGET_SUFFIX: loader_params[LoaderParamsKW.TARGET_SUFFIX],
+                           ConfigKW.UNDO_TRANSFORMS: undo_transforms,
+                           LoaderParamsKW.SLICE_AXIS: loader_params[LoaderParamsKW.SLICE_AXIS]})
 
     if command == "train":
         imed_utils.display_selected_transfoms(transform_train_params, dataset_type=["training"])
@@ -392,7 +442,7 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
         metric_fns = imed_metrics.get_metric_fns(ds_train.task)
 
         # If FiLM, normalize data
-        if 'film_layers' in model_params and any(model_params['film_layers']):
+        if ModelParamsKW.FILM_LAYERS in model_params and any(model_params[ModelParamsKW.FILM_LAYERS]):
             model_params, ds_train, ds_valid, train_onehotencoder = \
                 film_normalize_data(context, model_params, ds_train, ds_valid, path_output)
         else:
@@ -408,14 +458,15 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
             model_params=model_params,
             dataset_train=ds_train,
             dataset_val=ds_valid,
-            training_params=context["training_parameters"],
+            training_params=context[ConfigKW.TRAINING_PARAMETERS],
+            wandb_params=context.get(ConfigKW.WANDB),
             path_output=path_output,
             device=device,
             cuda_available=cuda_available,
             metric_fns=metric_fns,
             n_gif=n_gif,
             resume_training=resume_training,
-            debugging=context["debugging"])
+            debugging=context[ConfigKW.DEBUGGING])
 
     if thr_increment:
         # LOAD DATASET
@@ -429,21 +480,26 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
                                'training')
 
         # Choice of optimisation metric
-        metric = "recall_specificity" if model_params["name"] in imed_utils.CLASSIFIER_LIST else "dice"
+        if model_params[ModelParamsKW.NAME] in imed_utils.CLASSIFIER_LIST:
+            metric = MetricsKW.RECALL_SPECIFICITY
+        else:
+            metric = MetricsKW.DICE
+
         # Model path
-        model_path = os.path.join(path_output, "best_model.pt")
+        model_path = Path(path_output, "best_model.pt")
+
         # Run analysis
-        thr = imed_testing.threshold_analysis(model_path=model_path,
+        thr = imed_testing.threshold_analysis(model_path=str(model_path),
                                               ds_lst=[ds_train, ds_valid],
                                               model_params=model_params,
                                               testing_params=testing_params,
                                               metric=metric,
                                               increment=thr_increment,
-                                              fname_out=os.path.join(path_output, "roc.png"),
+                                              fname_out=str(Path(path_output, "roc.png")),
                                               cuda_available=cuda_available)
 
         # Update threshold in config file
-        context["postprocessing"]["binarize_prediction"] = {"thr": thr}
+        context[ConfigKW.POSTPROCESSING][PostprocessingKW.BINARIZE_PREDICTION] = {BinarizeProdictionKW.THR: thr}
         save_config_file(context, path_output)
 
     if command == 'train':
@@ -452,7 +508,7 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
     if command == 'test':
         # LOAD DATASET
         # Warn user that the input-level dropout is set during inference
-        if loader_params['is_input_dropout']:
+        if loader_params[LoaderParamsKW.IS_INPUT_DROPOUT]:
             logger.warning("Input-level dropout is set during testing. To turn this option off, set 'is_input_dropout'"
                            "to 'false' in the configuration file.")
         ds_test = imed_loader.load_dataset(bids_df, **{**loader_params, **{'data_list': test_lst,
@@ -461,9 +517,10 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
                                                                            'requires_undo': True}}, device=device,
                                            cuda_available=cuda_available)
 
-        metric_fns = imed_metrics.get_metric_fns(ds_test.task)
+        eval_params = context[ConfigKW.EVALUATION_PARAMETERS]
+        metric_fns = imed_metrics.get_metric_fns(ds_test.task, eval_params)
 
-        if 'film_layers' in model_params and any(model_params['film_layers']):
+        if ModelParamsKW.FILM_LAYERS in model_params and any(model_params[ModelParamsKW.FILM_LAYERS]):
             ds_test, model_params = update_film_model_params(context, ds_test, model_params, path_output)
 
         # RUN INFERENCE
@@ -474,17 +531,17 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
                                          device=device,
                                          cuda_available=cuda_available,
                                          metric_fns=metric_fns,
-                                         postprocessing=context['postprocessing'])
+                                         postprocessing=context[ConfigKW.POSTPROCESSING])
 
         # RUN EVALUATION
         df_results = imed_evaluation.evaluate(bids_df, path_output=path_output,
-                                              target_suffix=loader_params["target_suffix"],
-                                              eval_params=context["evaluation_parameters"])
+                                              target_suffix=loader_params[LoaderParamsKW.TARGET_SUFFIX],
+                                              eval_params=eval_params)
         return df_results, pred_metrics
 
 
 def create_dataset_and_ivadomed_version_log(context):
-    path_data = context['loader_parameters']['path_data']
+    path_data = context.get(ConfigKW.LOADER_PARAMETERS).get(LoaderParamsKW.PATH_DATA)
 
     ivadomed_version = imed_utils._version_string()
     datasets_version = []
@@ -495,12 +552,12 @@ def create_dataset_and_ivadomed_version_log(context):
         for Dataset in path_data:
             datasets_version.append(imed_utils.__get_commit(path_to_git_folder=Dataset))
 
-    log_file = os.path.join(context['path_output'], 'version_info.log')
+    path_log = Path(context.get(ConfigKW.PATH_OUTPUT), 'version_info.log')
 
     try:
-        f = open(log_file, "w")
+        f = path_log.open(mode="w")
     except OSError as err:
-        logger.error("OS error: {0}".format(err))
+        logger.error(f"OS error: {err}")
         raise Exception("Have you selected a log folder, and do you have write permissions for that folder?")
 
     # IVADOMED
@@ -558,15 +615,17 @@ def run_main():
     path_config_file = args.config
     context = imed_config_manager.ConfigurationManager(path_config_file).get_config()
 
-    context["command"] = imed_utils.get_command(args, context)
-    context["path_output"] = imed_utils.get_path_output(args, context)
-    context["loader_parameters"]["path_data"] = imed_utils.get_path_data(args, context)
+    context[ConfigKW.COMMAND] = imed_utils.get_command(args, context)
+    context[ConfigKW.PATH_OUTPUT] = imed_utils.get_path_output(args, context)
+    context[ConfigKW.LOADER_PARAMETERS][LoaderParamsKW.PATH_DATA] = imed_utils.get_path_data(args, context)
 
     # Run command
     run_command(context=context,
                 n_gif=args.gif if args.gif is not None else 0,
                 thr_increment=args.thr_increment if args.thr_increment else None,
-                resume_training=bool(args.resume_training))
+                resume_training=bool(args.resume_training),
+                no_patch=bool(args.no_patch),
+                overlap_2d=args.overlap_2d if args.overlap_2d else None)
 
 
 if __name__ == "__main__":
