@@ -6,7 +6,7 @@ from scipy import spatial
 
 
 # METRICS
-def get_metric_fns(task):
+def get_metric_fns(task, eval_params=None):
     metric_fns = [dice_score,
                   multi_class_dice_score,
                   precision_score,
@@ -14,7 +14,11 @@ def get_metric_fns(task):
                   specificity_score,
                   intersection_over_union,
                   accuracy_score]
-    if task == "segmentation":
+    if eval_params:
+        object_detection_metrics = eval_params['object_detection_metrics']
+    else:
+        object_detection_metrics = True
+    if task == "segmentation" and object_detection_metrics:
         metric_fns = metric_fns + [hausdorff_score]
 
     return metric_fns
@@ -81,10 +85,10 @@ def numeric_score(prediction, groundtruth):
     Returns:
         float, float, float, float: FP, FN, TP, TN
     """
-    FP = np.float(np.sum(prediction * (1.0 - groundtruth)))
-    FN = np.float(np.sum((1.0 - prediction) * groundtruth))
-    TP = np.float(np.sum(prediction * groundtruth))
-    TN = np.float(np.sum((1.0 - prediction) * (1.0 - groundtruth)))
+    FP = float(np.sum(prediction * (1.0 - groundtruth)))
+    FN = float(np.sum((1.0 - prediction) * groundtruth))
+    TP = float(np.sum(prediction * groundtruth))
+    TN = float(np.sum((1.0 - prediction) * (1.0 - groundtruth)))
     return FP, FN, TP, TN
 
 
