@@ -1,13 +1,11 @@
-import os
-
 from collections import OrderedDict
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 from torch.nn import Module
 from torch.nn import init
+from pathlib import Path
 import torchvision.models
-
 
 
 #Modified from torchvision.models.resnet.Resnet
@@ -142,14 +140,14 @@ class DenseNet(nn.Module):
         num_init_features (int) - the number of filters to learn in the first convolution layer
         bn_size (int) - multiplicative factor for number of bottle neck layers
           (i.e. bn_size * k features in the bottleneck layer)
-        drop_rate (float) - dropout rate after each dense layer
+        dropout_rate (float) - dropout rate after each dense layer
         num_classes (int) - number of classification classes
         memory_efficient (bool) - If True, uses checkpointing. Much more memory efficient,
           but slower. Default: *False*. See `"article" <https://arxiv.org/pdf/1707.06990.pdf>`_
     """
 
     def __init__(self, growth_rate=32, block_config=(6, 12, 24, 16),
-                 num_init_features=64, bn_size=4, drop_rate=0, num_classes=2, memory_efficient=False):
+                 num_init_features=64, bn_size=4, dropout_rate=0.3, num_classes=2, memory_efficient=False):
 
         super(DenseNet, self).__init__()
 
@@ -170,7 +168,7 @@ class DenseNet(nn.Module):
                 num_input_features=num_features,
                 bn_size=bn_size,
                 growth_rate=growth_rate,
-                drop_rate=drop_rate,
+                drop_rate=dropout_rate,
                 memory_efficient=memory_efficient
             )
             self.features.add_module('denseblock%d' % (i + 1), block)
@@ -215,14 +213,15 @@ class densenet121(DenseNet):
 
 
 class DownConv(Module):
-    """Two successive series of down convolution, batch normalization and drop out in 2D.
+    """Two successive series of down convolution, batch normalization and dropout in 2D.
     Used in U-Net's encoder.
 
     Args:
         in_feat (int): Number of channels in the input image.
         out_feat (int): Number of channels in the output image.
-        drop_rate (float): Probability of dropout.
+        dropout_rate (float): Probability of dropout.
         bn_momentum (float): Batch normalization momentum.
+        is_2d (bool): Indicates dimensionality of model: True for 2D convolutions, False for 3D convolutions.
 
     Attributes:
         conv1 (Conv2d): First 2D down convolution with kernel size 3 and padding of 1.
@@ -233,7 +232,7 @@ class DownConv(Module):
         conv2_drop (Dropout2d): Second 2D dropout.
     """
 
-    def __init__(self, in_feat, out_feat, drop_rate=0.4, bn_momentum=0.1, is_2d=True):
+    def __init__(self, in_feat, out_feat, dropout_rate=0.3, bn_momentum=0.1, is_2d=True):
         super(DownConv, self).__init__()
         if is_2d:
             conv = nn.Conv2d
@@ -246,11 +245,11 @@ class DownConv(Module):
 
         self.conv1 = conv(in_feat, out_feat, kernel_size=3, padding=1)
         self.conv1_bn = bn(out_feat, momentum=bn_momentum)
-        self.conv1_drop = dropout(drop_rate)
+        self.conv1_drop = dropout(dropout_rate)
 
         self.conv2 = conv(out_feat, out_feat, kernel_size=3, padding=1)
         self.conv2_bn = bn(out_feat, momentum=bn_momentum)
-        self.conv2_drop = dropout(drop_rate)
+        self.conv2_drop = dropout(dropout_rate)
 
     def forward(self, x):
         x = F.relu(self.conv1(x))
@@ -270,17 +269,18 @@ class UpConv(Module):
     Args:
         in_feat (int): Number of channels in the input image.
         out_feat (int): Number of channels in the output image.
-        drop_rate (float): Probability of dropout.
+        dropout_rate (float): Probability of dropout.
         bn_momentum (float): Batch normalization momentum.
+        is_2d (bool): Indicates dimensionality of model: True for 2D convolutions, False for 3D convolutions.
 
     Attributes:
         downconv (DownConv): Down convolution.
     """
 
-    def __init__(self, in_feat, out_feat, drop_rate=0.4, bn_momentum=0.1, is_2d=True):
+    def __init__(self, in_feat, out_feat, dropout_rate=0.3, bn_momentum=0.1, is_2d=True):
         super(UpConv, self).__init__()
         self.is_2d = is_2d
-        self.downconv = DownConv(in_feat, out_feat, drop_rate, bn_momentum, is_2d)
+        self.downconv = DownConv(in_feat, out_feat, dropout_rate, bn_momentum, is_2d)
 
     def forward(self, x, y):
         # For retrocompatibility purposes
@@ -301,10 +301,12 @@ class Encoder(Module):
     Args:
         in_channel (int): Number of channels in the input image.
         depth (int): Number of down convolutions minus bottom down convolution.
-        drop_rate (float): Probability of dropout.
+        dropout_rate (float): Probability of dropout.
         bn_momentum (float): Batch normalization momentum.
         n_metadata (dict): FiLM metadata see ivadomed.loader.film for more details.
         film_layers (list): List of 0 or 1 indicating on which layer FiLM is applied.
+        is_2d (bool): Indicates dimensionality of model: True for 2D convolutions, False for 3D convolutions.
+        n_filters (int):  Number of base filters in the U-Net.
 
     Attributes:
         depth (int): Number of down convolutions minus bottom down convolution.
@@ -313,25 +315,25 @@ class Encoder(Module):
         film_bottom (FiLMlayer): FiLM layer applied to bottom convolution.
     """
 
-    def __init__(self, in_channel=1, depth=3, drop_rate=0.4, bn_momentum=0.1, n_metadata=None, film_layers=None,
+    def __init__(self, in_channel=1, depth=3, dropout_rate=0.3, bn_momentum=0.1, n_metadata=None, film_layers=None,
                  is_2d=True, n_filters=64):
         super(Encoder, self).__init__()
         self.depth = depth
         self.down_path = nn.ModuleList()
         # first block
-        self.down_path.append(DownConv(in_channel, n_filters, drop_rate, bn_momentum, is_2d))
+        self.down_path.append(DownConv(in_channel, n_filters, dropout_rate, bn_momentum, is_2d))
         self.down_path.append(FiLMlayer(n_metadata, n_filters) if film_layers and film_layers[0] else None)
         max_pool = nn.MaxPool2d if is_2d else nn.MaxPool3d
         self.down_path.append(max_pool(2))
 
         for i in range(depth - 1):
-            self.down_path.append(DownConv(n_filters, n_filters * 2, drop_rate, bn_momentum, is_2d))
+            self.down_path.append(DownConv(n_filters, n_filters * 2, dropout_rate, bn_momentum, is_2d))
             self.down_path.append(FiLMlayer(n_metadata, n_filters * 2) if film_layers and film_layers[i + 1] else None)
             self.down_path.append(max_pool(2))
             n_filters = n_filters * 2
 
         # Bottom
-        self.conv_bottom = DownConv(n_filters, n_filters, drop_rate, bn_momentum, is_2d)
+        self.conv_bottom = DownConv(n_filters, n_filters, dropout_rate, bn_momentum, is_2d)
         self.film_bottom = FiLMlayer(n_metadata, n_filters) if film_layers and film_layers[self.depth] else None
 
     def forward(self, x, context=None):
@@ -367,12 +369,14 @@ class Decoder(Module):
     Args:
         out_channel (int): Number of channels in the output image.
         depth (int): Number of down convolutions minus bottom down convolution.
-        drop_rate (float): Probability of dropout.
+        dropout_rate (float): Probability of dropout.
         bn_momentum (float): Batch normalization momentum.
         n_metadata (dict): FiLM metadata see ivadomed.loader.film for more details.
         film_layers (list): List of 0 or 1 indicating on which layer FiLM is applied.
         hemis (bool): Boolean indicating if HeMIS is on or not.
-        final_activation (str): Choice of final activation between "sigmoid", "relu" and "softmax"
+        final_activation (str): Choice of final activation between "sigmoid", "relu" and "softmax".
+        is_2d (bool): Indicates dimensionality of model: True for 2D convolutions, False for 3D convolutions.
+        n_filters (int):  Number of base filters in the U-Net.
 
     Attributes:
         depth (int): Number of down convolutions minus bottom down convolution.
@@ -383,7 +387,7 @@ class Decoder(Module):
         softmax (Softmax): Softmax layer that can be applied as last layer.
     """
 
-    def __init__(self, out_channel=1, depth=3, drop_rate=0.4, bn_momentum=0.1,
+    def __init__(self, out_channel=1, depth=3, dropout_rate=0.3, bn_momentum=0.1,
                  n_metadata=None, film_layers=None, hemis=False, final_activation="sigmoid", is_2d=True, n_filters=64):
         super(Decoder, self).__init__()
         self.depth = depth
@@ -393,7 +397,7 @@ class Decoder(Module):
         self.up_path = nn.ModuleList()
         if hemis:
             in_channel = n_filters * 2 ** self.depth
-            self.up_path.append(UpConv(in_channel * 2, n_filters * 2 ** (self.depth - 1), drop_rate, bn_momentum,
+            self.up_path.append(UpConv(in_channel * 2, n_filters * 2 ** (self.depth - 1), dropout_rate, bn_momentum,
                                        is_2d))
             if film_layers and film_layers[self.depth + 1]:
                 self.up_path.append(FiLMlayer(n_metadata, n_filters * 2 ** (self.depth - 1)))
@@ -403,7 +407,7 @@ class Decoder(Module):
         else:
             in_channel = n_filters * 2 ** self.depth
 
-            self.up_path.append(UpConv(in_channel, n_filters * 2 ** (self.depth - 1), drop_rate, bn_momentum, is_2d))
+            self.up_path.append(UpConv(in_channel, n_filters * 2 ** (self.depth - 1), dropout_rate, bn_momentum, is_2d))
             if film_layers and film_layers[self.depth + 1]:
                 self.up_path.append(FiLMlayer(n_metadata, n_filters * 2 ** (self.depth - 1)))
             else:
@@ -414,7 +418,7 @@ class Decoder(Module):
 
             self.up_path.append(
                 UpConv(in_channel + n_filters * 2 ** (self.depth - i - 1 + int(hemis)), n_filters * 2 ** (self.depth - i - 1),
-                       drop_rate, bn_momentum, is_2d))
+                       dropout_rate, bn_momentum, is_2d))
             if film_layers and film_layers[self.depth + i + 1]:
                 self.up_path.append(FiLMlayer(n_metadata, n_filters * 2 ** (self.depth - i - 1)))
             else:
@@ -439,7 +443,9 @@ class Decoder(Module):
         if self.last_film:
             x, w_film = self.last_film(x, context, w_film)
 
-        if hasattr(self, "final_activation") and self.final_activation == "softmax":
+        if hasattr(self, "final_activation") and self.final_activation not in ["softmax", "relu", "sigmoid"]:
+            raise ValueError("final_activation value has to be either softmax, relu, or sigmoid")
+        elif hasattr(self, "final_activation") and self.final_activation == "softmax":
             preds = self.softmax(x)
         elif hasattr(self, "final_activation") and self.final_activation == "relu":
             preds = nn.ReLU()(x) / nn.ReLU()(x).max()
@@ -475,7 +481,7 @@ class Unet(Module):
         in_channel (int): Number of channels in the input image.
         out_channel (int): Number of channels in the output image.
         depth (int): Number of down convolutions minus bottom down convolution.
-        drop_rate (float): Probability of dropout.
+        dropout_rate (float): Probability of dropout.
         bn_momentum (float): Batch normalization momentum.
         final_activation (str): Choice of final activation between "sigmoid", "relu" and "softmax".
         is_2d (bool): Indicates dimensionality of model: True for 2D convolutions, False for 3D convolutions.
@@ -487,16 +493,16 @@ class Unet(Module):
         decoder (Decoder): U-net decoder.
     """
 
-    def __init__(self, in_channel=1, out_channel=1, depth=3, drop_rate=0.4, bn_momentum=0.1, final_activation='sigmoid',
+    def __init__(self, in_channel=1, out_channel=1, depth=3, dropout_rate=0.3, bn_momentum=0.1, final_activation='sigmoid',
                  is_2d=True, n_filters=64, **kwargs):
         super(Unet, self).__init__()
 
         # Encoder path
-        self.encoder = Encoder(in_channel=in_channel, depth=depth, drop_rate=drop_rate, bn_momentum=bn_momentum,
+        self.encoder = Encoder(in_channel=in_channel, depth=depth, dropout_rate=dropout_rate, bn_momentum=bn_momentum,
                                is_2d=is_2d, n_filters=n_filters)
 
         # Decoder path
-        self.decoder = Decoder(out_channel=out_channel, depth=depth, drop_rate=drop_rate, bn_momentum=bn_momentum,
+        self.decoder = Decoder(out_channel=out_channel, depth=depth, dropout_rate=dropout_rate, bn_momentum=bn_momentum,
                                final_activation=final_activation, is_2d=is_2d, n_filters=n_filters)
 
     def forward(self, x):
@@ -513,7 +519,7 @@ class FiLMedUnet(Unet):
         n_channel (int): Number of channels in the input image.
         out_channel (int): Number of channels in the output image.
         depth (int): Number of down convolutions minus bottom down convolution.
-        drop_rate (float): Probability of dropout.
+        dropout_rate (float): Probability of dropout.
         bn_momentum (float): Batch normalization momentum.
         n_metadata (dict): FiLM metadata see ivadomed.loader.film for more details.
         film_layers (list): List of 0 or 1 indicating on which layer FiLM is applied.
@@ -526,9 +532,10 @@ class FiLMedUnet(Unet):
         decoder (Decoder): U-net decoder.
     """
 
-    def __init__(self, in_channel=1, out_channel=1, depth=3, drop_rate=0.4,
+    def __init__(self, in_channel=1, out_channel=1, depth=3, dropout_rate=0.3,
                  bn_momentum=0.1, n_metadata=None, film_layers=None, is_2d=True, n_filters=64, **kwargs):
-        super().__init__(in_channel=1, out_channel=1, depth=3, drop_rate=0.4, bn_momentum=0.1)
+        super().__init__(in_channel=in_channel, out_channel=out_channel, depth=depth,
+                         dropout_rate=dropout_rate, bn_momentum=bn_momentum)
 
         # Verify if the length of boolean FiLM layers corresponds to the depth
         if film_layers:
@@ -538,10 +545,10 @@ class FiLMedUnet(Unet):
         else:
             film_layers = [0] * (2 * depth + 2)
         # Encoder path
-        self.encoder = Encoder(in_channel=in_channel, depth=depth, drop_rate=drop_rate, bn_momentum=bn_momentum,
+        self.encoder = Encoder(in_channel=in_channel, depth=depth, dropout_rate=dropout_rate, bn_momentum=bn_momentum,
                                n_metadata=n_metadata, film_layers=film_layers, is_2d=is_2d, n_filters=n_filters)
         # Decoder path
-        self.decoder = Decoder(out_channel=out_channel, depth=depth, drop_rate=drop_rate, bn_momentum=bn_momentum,
+        self.decoder = Decoder(out_channel=out_channel, depth=depth, dropout_rate=dropout_rate, bn_momentum=bn_momentum,
                                n_metadata=n_metadata, film_layers=film_layers, is_2d=is_2d, n_filters=n_filters)
 
     def forward(self, x, context=None):
@@ -680,7 +687,7 @@ class HeMISUnet(Module):
         contrasts (list): List of contrasts.
         out_channel (int): Number of output channels.
         depth (int): Number of down convolutions minus bottom down convolution.
-        drop_rate (float): Probability of dropout.
+        dropout_rate (float): Probability of dropout.
         bn_momentum (float): Batch normalization momentum.
         **kwargs:
 
@@ -691,18 +698,18 @@ class HeMISUnet(Module):
         decoder (Decoder): U-Net decoder.
     """
 
-    def __init__(self, contrasts, out_channel=1, depth=3, drop_rate=0.4, bn_momentum=0.1, **kwargs):
+    def __init__(self, contrasts, out_channel=1, depth=3, dropout_rate=0.3, bn_momentum=0.1, **kwargs):
         super(HeMISUnet, self).__init__()
         self.depth = depth
         self.contrasts = contrasts
 
         # Encoder path
         self.Encoder_mod = nn.ModuleDict(
-            [['Encoder_{}'.format(Mod), Encoder(in_channel=1, depth=depth, drop_rate=drop_rate,
+            [['Encoder_{}'.format(Mod), Encoder(in_channel=1, depth=depth, dropout_rate=dropout_rate,
                                                 bn_momentum=bn_momentum)] for Mod in self.contrasts])
 
         # Decoder path
-        self.decoder = Decoder(out_channel=out_channel, depth=depth, drop_rate=drop_rate,
+        self.decoder = Decoder(out_channel=out_channel, depth=depth, dropout_rate=dropout_rate,
                                bn_momentum=bn_momentum, hemis=True)
 
     def forward(self, x_mods, indexes_mod):
@@ -750,7 +757,7 @@ class Modified3DUNet(nn.Module):
         out_channel (int): Number of channels in the output image.
         n_filters (int): Number of base filters in the U-Net.
         attention (bool): Boolean indicating whether the attention module is on or not.
-        drop_rate (float): Probability of dropout.
+        dropout_rate (float): Probability of dropout.
         bn_momentum (float): Batch normalization momentum.
         final_activation (str): Choice of final activation between "sigmoid", "relu" and "softmax".
         **kwargs:
@@ -766,7 +773,7 @@ class Modified3DUNet(nn.Module):
     Note: All layers are defined as attributes and used in the forward method.
     """
 
-    def __init__(self, in_channel, out_channel, n_filters=16, attention=False, drop_rate=0.6, bn_momentum=0.1,
+    def __init__(self, in_channel, out_channel, n_filters=16, attention=False, dropout_rate=0.3, bn_momentum=0.1,
                  final_activation="sigmoid", n_metadata=None, film_layers=None, **kwargs):
         super(Modified3DUNet, self).__init__()
         self.in_channels = in_channel
@@ -777,7 +784,7 @@ class Modified3DUNet(nn.Module):
         self.final_activation = final_activation
 
         self.lrelu = nn.LeakyReLU()
-        self.dropout3d = nn.Dropout3d(p=drop_rate)
+        self.dropout3d = nn.Dropout3d(p=dropout_rate)
         self.upsacle = nn.Upsample(scale_factor=2, mode='nearest')
         self.softmax = nn.Softmax(dim=1)
 
@@ -1107,7 +1114,9 @@ class Modified3DUNet(nn.Module):
             out, w_film = self.film_layer10(out, context, w_film)
         seg_layer = out
 
-        if hasattr(self, "final_activation") and self.final_activation == "softmax":
+        if hasattr(self, "final_activation") and self.final_activation not in ["softmax", "relu", "sigmoid"]:
+            raise ValueError("final_activation value has to be either softmax, relu, or sigmoid")
+        elif hasattr(self, "final_activation") and self.final_activation == "softmax":
             out = self.softmax(out)
         elif hasattr(self, "final_activation") and self.final_activation == "relu":
             out = nn.ReLU()(seg_layer) / nn.ReLU()(seg_layer).max() if bool(nn.ReLU()(seg_layer).max()) \
@@ -1126,11 +1135,11 @@ class UNet3D(Modified3DUNet):
     """To ensure retrocompatibility, when calling UNet3D (old model name), Modified3DUNet will be called.
     see Modified3DUNet to learn more about parameters.
     """
-    def __init__(self, in_channel, out_channel, n_filters=16, attention=False, drop_rate=0.6, bn_momentum=0.1,
+    def __init__(self, in_channel, out_channel, n_filters=16, attention=False, dropout_rate=0.3, bn_momentum=0.1,
                  final_activation="sigmoid", n_metadata=None, film_layers=None, **kwargs):
         super(UNet3D, self).__init__()
         Modified3DUNet(in_channel=in_channel, out_channel=out_channel, n_filters=n_filters, attention=attention,
-                       drop_rate=drop_rate, bn_momentum=bn_momentum, final_activation=final_activation,
+                       dropout_rate=dropout_rate, bn_momentum=bn_momentum, final_activation=final_activation,
                        n_metadata=n_metadata, film_layers=film_layers, **kwargs)
 
 
@@ -1476,27 +1485,38 @@ def set_model_for_retrain(model_path, retrain_fraction, map_location, reset=True
 def get_model_filenames(folder_model):
     """Get trained model filenames from its folder path.
 
-    This function checks if the folder_model exists and get trained model (.pt or .onnx) and its configuration file
-    (.json) from it.
-    Note: if the model exists as .onnx, then this function returns its onnx path instead of the .pt version.
+    This function checks if the folder_model exists and get trained model path (.pt or .onnx based on
+    model and GPU availability) and its configuration file (.json) from it.
 
     Args:
         folder_name (str): Path of the model folder.
 
     Returns:
-        str, str: Paths of the model (.onnx) and its configuration file (.json).
+        str, str: Paths of the model (.pt or .onnx) and its configuration file (.json).
     """
-    if os.path.isdir(folder_model):
-        prefix_model = os.path.basename(folder_model)
-        # Check if model and model metadata exist. Verify if ONNX model exists, if not try to find .pt model
-        fname_model = os.path.join(folder_model, prefix_model + '.onnx')
-        if not os.path.isfile(fname_model):
-            fname_model = os.path.join(folder_model, prefix_model + '.pt')
-            if not os.path.exists(fname_model):
-                raise FileNotFoundError(fname_model)
-        fname_model_metadata = os.path.join(folder_model, prefix_model + '.json')
-        if not os.path.isfile(fname_model_metadata):
-            raise FileNotFoundError(fname_model)
+    if Path(folder_model).is_dir():
+        prefix_model = Path(folder_model).name
+        fname_model_onnx = Path(folder_model, prefix_model + '.onnx')
+        fname_model_pt = Path(folder_model, prefix_model + '.pt')
+        cuda_available = torch.cuda.is_available()
+
+        # Assign '.pt' or '.onnx' model based on file existence and GPU/CPU device availability
+        if not fname_model_pt.is_file() and not fname_model_onnx.is_file():
+            raise FileNotFoundError(f"Model files not found in model folder: "
+                                    f"'{str(fname_model_onnx)}' or '{str(fname_model_pt)}'")
+        # '.pt' is preferred on GPU, or on CPU if '.onnx' doesn't exist
+        elif ((    cuda_available and     fname_model_pt.is_file()) or
+              (not cuda_available and not fname_model_onnx.is_file())):
+            fname_model = fname_model_pt
+        # '.onnx' is preferred on CPU, or on GPU if '.pt' doesn't exist
+        elif ((not cuda_available and     fname_model_onnx.is_file()) or
+              (    cuda_available and not fname_model_pt.is_file())):
+            fname_model = fname_model_onnx
+
+        fname_model_metadata = Path(folder_model, prefix_model + '.json')
+        if not fname_model_metadata.is_file():
+            raise FileNotFoundError(f"Model config file not found in model folder: '{str(fname_model_metadata)}'")
     else:
-        raise FileNotFoundError(fname_model)
-    return fname_model, fname_model_metadata
+        raise FileNotFoundError(folder_model)
+
+    return str(fname_model), str(fname_model_metadata)
