@@ -1,14 +1,13 @@
 # Deals with postprocessing on generated segmentation.
 
 import functools
-import os
 
 import nibabel as nib
 import numpy as np
 from loguru import logger
-from scipy.ndimage import label, generate_binary_structure
-from scipy.ndimage.morphology import binary_fill_holes
+from scipy.ndimage import label, generate_binary_structure, binary_fill_holes
 from skimage.feature import peak_local_max
+from pathlib import Path
 
 
 def nifti_capable(wrapped):
@@ -24,7 +23,11 @@ def nifti_capable(wrapped):
     @functools.wraps(wrapped)
     def wrapper(data, *args, **kwargs):
         if isinstance(data, nib.Nifti1Image):
-            return nib.Nifti1Image(wrapper(np.copy(np.asanyarray(data.dataobj)), *args, **kwargs), data.affine)
+            return nib.Nifti1Image(
+                dataobj=wrapper(np.copy(np.asanyarray(data.dataobj)), *args, **kwargs),
+                affine=data.header.get_best_affine(),
+                header=data.header.copy()
+            )
         return wrapped(data, *args, **kwargs)
 
     return wrapper
@@ -90,7 +93,7 @@ def threshold_predictions(predictions, thr=0.5):
     thresholded_preds[low_values_indices] = 0
     low_values_indices = thresholded_preds >= thr
     thresholded_preds[low_values_indices] = 1
-    return thresholded_preds.astype(np.int)
+    return thresholded_preds.astype(int)
 
 
 @nifti_capable
@@ -151,7 +154,7 @@ def fill_holes(predictions, structure=(3, 3, 3)):
     """
     assert np.array_equal(predictions, predictions.astype(bool))
     assert len(structure) == len(predictions.shape)
-    return binary_fill_holes(predictions, structure=np.ones(structure)).astype(np.int)
+    return binary_fill_holes(predictions, structure=np.ones(structure)).astype(int)
 
 
 @nifti_capable
@@ -210,7 +213,10 @@ def label_file_from_coordinates(nifti_image, coord_list):
     for j in range(len(coord_list)):
         label_array[coord_list[j][0], coord_list[j][1], coord_list[j][2]] = 1
 
-    nib_pred = nib.Nifti1Image(label_array, nifti_image.affine)
+    nib_pred = nib.Nifti1Image(
+        dataobj=label_array,
+        affine=nifti_image.header.get_best_affine(),
+    )
 
     return nib_pred
 
@@ -230,7 +236,7 @@ def remove_small_objects(data, bin_structure, size_min):
     data_label, n = label(data, structure=bin_structure)
 
     for idx in range(1, n + 1):
-        data_idx = (data_label == idx).astype(np.int)
+        data_idx = (data_label == idx).astype(int)
         n_nonzero = np.count_nonzero(data_idx)
 
         if n_nonzero < size_min:
@@ -309,7 +315,7 @@ class Postprocessing(object):
         """
         if thr >= 0:
             uncertainty_path = self.filename_prefix + suffix
-            if os.path.exists(uncertainty_path):
+            if Path(uncertainty_path).exists():
                 data_uncertainty = nib.load(uncertainty_path).get_fdata()
                 if suffix == "_unc-iou.nii.gz" or suffix == "_soft.nii.gz":
                     self.data_pred = mask_predictions(self.data_pred, data_uncertainty > thr)
