@@ -1,17 +1,20 @@
 #!/usr/bin/env python
 
-import os
 import argparse
 import nibabel as nib
 import numpy as np
 import random
 import torch
 
+from pathlib import Path
+from loguru import logger
 from ivadomed import config_manager as imed_config_manager
 from ivadomed.loader import utils as imed_loader_utils
+from ivadomed.loader.sample_meta_data import SampleMetadata
 from ivadomed import transforms as imed_transforms
 from ivadomed import utils as imed_utils
 from ivadomed import maths as imed_maths
+from ivadomed.keywords import ConfigKW, TransformationKW, LoaderParamsKW, MetadataKW
 
 
 def get_parser():
@@ -97,11 +100,11 @@ def run_visualization(input, config, number, output, roi):
     context = imed_config_manager.ConfigurationManager(config).get_config()
 
     # Create output folder
-    if not os.path.isdir(output):
-        os.makedirs(output)
+    if not Path(output).is_dir():
+        Path(output).mkdir(parents=True)
 
     # Slice extracted according to below axis
-    axis = imed_utils.AXIS_DCT[context["loader_parameters"]["slice_axis"]]
+    axis = imed_utils.AXIS_DCT[context[ConfigKW.LOADER_PARAMETERS][LoaderParamsKW.SLICE_AXIS]]
     # Get data
     input_img, input_data = get_data(input, axis)
     # Image or Mask
@@ -112,10 +115,10 @@ def run_visualization(input, config, number, output, roi):
     indexes = random.sample(range(0, input_data.shape[2]), number)
 
     # Get training transforms
-    training_transforms, _, _ = imed_transforms.get_subdatasets_transforms(context["transformation"])
+    training_transforms, _, _ = imed_transforms.get_subdatasets_transforms(context[ConfigKW.TRANSFORMATION])
 
-    if "ROICrop" in training_transforms:
-        if roi and os.path.isfile(roi):
+    if TransformationKW.ROICROP in training_transforms:
+        if roi and Path(roi).is_file():
             roi_img, roi_data = get_data(roi, axis)
         else:
             raise ValueError("\nPlease provide ROI image (-r) in order to apply ROICrop transformation.")
@@ -139,11 +142,12 @@ def run_visualization(input, config, number, output, roi):
         for i in indexes:
             data = [input_data[:, :, i]]
             # Init metadata
-            metadata = imed_loader_utils.SampleMetadata({"zooms": zooms, "data_type": "gt" if is_mask else "im"})
+            metadata = SampleMetadata({MetadataKW.ZOOMS: zooms, MetadataKW.DATA_TYPE: "gt" if is_mask else "im"})
 
             # Apply transformations to ROI
-            if "CenterCrop" in training_transforms or ("ROICrop" in training_transforms and os.path.isfile(roi)):
-                metadata.__setitem__('crop_params', {})
+            if TransformationKW.CENTERCROP in training_transforms or \
+                    (TransformationKW.ROICROP in training_transforms and Path(roi).is_file()):
+                metadata.__setitem__(MetadataKW.CROP_PARAMS, {})
 
             # Apply transformations to image
             stack_im, _ = composed_transforms(sample=data,
@@ -151,9 +155,9 @@ def run_visualization(input, config, number, output, roi):
                                               data_type="im")
 
             # Plot before / after transformation
-            fname_out = os.path.join(output, stg_transforms+"slice"+str(i)+".png")
-            print("Fname out: {}.".format(fname_out))
-            print("\t{}".format(dict(metadata)))
+            fname_out = str(Path(output, stg_transforms + "slice" + str(i) + ".png"))
+            logger.debug(f"Fname out: {fname_out}.")
+            logger.debug(f"\t{dict(metadata)}")
             # rescale intensities
             if len(stg_transforms[:-1].split("_")) == 1:
                 before = np.rot90(imed_maths.rescale_values_array(data[0], 0.0, 1.0))
