@@ -2,10 +2,17 @@ import os
 import sys
 import subprocess
 import hashlib
+import datetime
+import platform
+
+import numpy as np
+import wandb
 from enum import Enum
 from loguru import logger
-
+from pathlib import Path
+from ivadomed.keywords import ConfigKW, LoaderParamsKW, WandbKW
 from typing import List
+from difflib import SequenceMatcher
 
 AXIS_DCT = {'sagittal': 0, 'coronal': 1, 'axial': 2}
 
@@ -25,6 +32,43 @@ class Metavar(Enum):
 
     def __str__(self):
         return self.value
+
+
+def initialize_wandb(wandb_params):
+    """Initializes WandB and based upon the parameters sets it up or disables it for experimental tracking
+    
+    Args:
+        wandb_params (dict): wandb parameters
+        
+    Returns:
+        bool, wandb_tracking: True if wandb tracking is enabled        
+    """
+    try:
+        # raise an error if the key is empty
+        if not bool(wandb_params[WandbKW.WANDB_API_KEY].strip()):
+            raise ValueError()
+
+        # Log on to WandB (assuming that the API Key is correct)
+        # if not, login would raise an exception for the cases invalid API key and not found
+        wandb.login(key=wandb_params[WandbKW.WANDB_API_KEY], anonymous='allow', timeout=60)
+
+    except Exception as e:
+        # log error mssg for unsuccessful wandb authentication
+        if wandb_params is not None:
+            logger.info("Incorrect WandB API Key! Please re-check the entered API key.")
+            logger.info("Disabling WandB Tracking, continuing with Tensorboard Logging")
+        else:
+            logger.info("No WandB parameters found! Continuing with Tensorboard Logging")
+
+        # set flag
+        wandb_tracking = False
+
+    else:
+        # setting flag after successful authentication
+        logger.info("WandB API Authentication Successful!")
+        wandb_tracking = True
+
+    return wandb_tracking
 
 
 def get_task(model_name):
@@ -78,7 +122,7 @@ def generate_sha_256(context: dict, df, file_lst: List[str]) -> None:
     assert isinstance(df, DataFrame)
 
     # generating sha256 for list of data
-    context['training_sha256'] = {}
+    context[ConfigKW.TRAINING_SHA256] = {}
     # file_list is a list of filename strings
     for file in file_lst:
         # bids_df is a dataframe with column values path...filename...
@@ -89,7 +133,7 @@ def generate_sha_256(context: dict, df, file_lst: List[str]) -> None:
         with open(file_path, "rb") as f:
             for byte_block in iter(lambda: f.read(4096), b""):
                 sha256_hash.update(byte_block)
-            context['training_sha256'][file] = sha256_hash.hexdigest()
+            context[ConfigKW.TRAINING_SHA256][file] = sha256_hash.hexdigest()
 
 
 def save_onnx_model(model, inputs, model_path):
@@ -177,6 +221,7 @@ def plot_transformed_sample(before, after, list_title=None, fname_out="", cmap="
         list_title = ['Sample before transform', 'Sample after transform']
 
     plt.interactive(False)
+    plt.rcParams.update({'figure.max_open_warning': 0})
     plt.figure(figsize=(20, 10))
 
     plt.subplot(1, 2, 1)
@@ -192,7 +237,6 @@ def plot_transformed_sample(before, after, list_title=None, fname_out="", cmap="
     if fname_out:
         plt.savefig(fname_out)
     else:
-        matplotlib.use('TkAgg')
         plt.show()
 
 
@@ -209,7 +253,7 @@ def _git_info(commit_env='IVADOMED_COMMIT', branch_env='IVADOMED_BRANCH'):
     """
     ivadomed_commit = os.getenv(commit_env, "unknown")
     ivadomed_branch = os.getenv(branch_env, "unknown")
-    if check_exe("git") and os.path.isdir(os.path.join(__ivadomed_dir__, ".git")):
+    if check_exe("git") and Path(__ivadomed_dir__, ".git").is_dir():
         ivadomed_commit = __get_commit() or ivadomed_commit
         ivadomed_branch = __get_branch() or ivadomed_branch
 
@@ -218,8 +262,8 @@ def _git_info(commit_env='IVADOMED_COMMIT', branch_env='IVADOMED_BRANCH'):
     else:
         install_type = 'package'
 
-    path_version = os.path.join(__ivadomed_dir__, 'ivadomed', 'version.txt')
-    with open(path_version) as f:
+    path_version = Path(__ivadomed_dir__, 'ivadomed', 'version.txt')
+    with path_version.open() as f:
         version_ivadomed = f.read().strip()
 
     return install_type, ivadomed_commit, ivadomed_branch, version_ivadomed
@@ -235,15 +279,15 @@ def check_exe(name):
     """
 
     def is_exe(fpath):
-        return os.path.isfile(fpath) and os.access(fpath, os.X_OK)
+        return Path(fpath).is_file() and os.access(fpath, os.X_OK)
 
-    fpath, fname = os.path.split(name)
+    fpath = Path(name).parent
     if fpath and is_exe(name):
         return fpath
     else:
         for path in os.environ["PATH"].split(os.pathsep):
             path = path.strip('"')
-            exe_file = os.path.join(path, name)
+            exe_file = str(Path(path, name))
             if is_exe(exe_file):
                 return exe_file
 
@@ -264,10 +308,7 @@ def get_arguments(parser, args):
             ["-d", "SOME_ARG", "--model", "SOME_ARG"]
     """
     try:
-        if args:
-            args = parser.parse_args(args)
-        else:
-            args = parser.parse_args(args=None if sys.argv[1:] else ['--help'])
+        args = parser.parse_args(args)
     except SystemExit as e:
         if e.code != 0:  # Calling `--help` raises SystemExit with 0 exit code (i.e. not an ArgParseException)
             raise ArgParseException('Error parsing args')
@@ -288,7 +329,7 @@ def __get_commit(path_to_git_folder=None):
     if path_to_git_folder is None:
         path_to_git_folder = __ivadomed_dir__
     else:
-        path_to_git_folder = os.path.abspath(os.path.expanduser(path_to_git_folder))
+        path_to_git_folder = Path(path_to_git_folder).expanduser().absolute()
 
     p = subprocess.Popen(["git", "rev-parse", "HEAD"], stdout=subprocess.PIPE, stderr=subprocess.PIPE,
                          cwd=path_to_git_folder)
@@ -343,7 +384,7 @@ def _version_string():
         return "{install_type}-{ivadomed_branch}-{ivadomed_commit}".format(**locals())
 
 
-__ivadomed_dir__ = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
+__ivadomed_dir__ = Path(__file__).resolve().parent.parent
 __version__ = _version_string()
 
 
@@ -355,11 +396,13 @@ def get_command(args, context):
     elif args.segment:
         return "segment"
     else:
-        logger.info("No CLI argument given for command: ( --train | --test | --segment ). Will check config file for command...")
+        logger.info(
+            "No CLI argument given for command: ( --train | --test | --segment ). Will check config file for command...")
 
         try:
-            if context["command"] == "train" or context["command"] == "test" or context["command"] == "segment":
-                return context["command"]
+            if context[ConfigKW.COMMAND] == "train" or context[ConfigKW.COMMAND] == "test" or context[
+                ConfigKW.COMMAND] == "segment":
+                return context[ConfigKW.COMMAND]
             else:
                 logger.error("Specified invalid command argument in config file.")
         except AttributeError:
@@ -370,10 +413,11 @@ def get_path_output(args, context):
     if args.path_output:
         return args.path_output
     else:
-        logger.info("CLI flag --path-output not used to specify output directory. Will check config file for directory...")
+        logger.info(
+            "CLI flag --path-output not used to specify output directory. Will check config file for directory...")
         try:
-            if context["path_output"]:
-                return context["path_output"]
+            if context[ConfigKW.PATH_OUTPUT]:
+                return context[ConfigKW.PATH_OUTPUT]
         except AttributeError:
             logger.error("Have not specified a path-output argument via CLI nor config file.")
 
@@ -382,10 +426,11 @@ def get_path_data(args, context):
     if args.path_data:
         return args.path_data
     else:
-        logger.info("CLI flag --path-data not used to specify BIDS data directory. Will check config file for directory...")
+        logger.info(
+            "CLI flag --path-data not used to specify BIDS data directory. Will check config file for directory...")
         try:
-            if context["loader_parameters"]["path_data"]:
-                return context["loader_parameters"]["path_data"]
+            if context[ConfigKW.LOADER_PARAMETERS][LoaderParamsKW.PATH_DATA]:
+                return context[ConfigKW.LOADER_PARAMETERS][LoaderParamsKW.PATH_DATA]
         except AttributeError:
             logger.error("Have not specified a path-data argument via CLI nor config file.")
 
@@ -404,7 +449,103 @@ def format_path_data(path_data):
     return path_data
 
 
+def similarity_score(a: str, b: str) -> float:
+    """
+    use DiffLIb SequenceMatcher to resolve the similarity between text. Help make better choice in terms of derivatives.
+    Args:
+        a: a string
+        b: another string
+    Returns: a score indicative of the similarity between the sequence.
+    """
+    return SequenceMatcher(None, a, b).ratio()
+
+
 def init_ivadomed():
     """Initialize the ivadomed for typical terminal usage."""
     # Display ivadomed version
     logger.info('\nivadomed ({})\n'.format(__version__))
+
+
+def print_stats(arr):
+    logger.info(f"\tMean: {np.mean(arr)} %")
+    logger.info(f"\tMedian: {np.median(arr)} %")
+    logger.info(f"\tInter-quartile range: [{np.percentile(arr, 25)}, {np.percentile(arr, 75)}] %")
+
+
+def get_timestamp() -> str:
+    """
+    Return a datetime string in the format YYYY-MM-DDTHHMMSS.(sub-precision)
+    Returns:
+    """
+    timestamp = datetime.datetime.now().isoformat().replace(":", "")
+    return timestamp
+
+
+def get_system_memory() -> float:
+    """
+    Return the system memory in GB.
+    Returns:
+    """
+    current_platform = platform.system()
+    if current_platform == "Linux":
+        return get_linux_system_memory()
+    elif current_platform == "Windows":
+        return get_win_system_memory()
+    elif current_platform == "Darwin":
+        return get_mac_system_memory()
+
+
+def get_win_system_memory() -> float:
+    """
+    Obtain the amount of memory available on Windows system.
+    Returns: memory in GB
+    Source: https://stackoverflow.com/a/21589439
+    """
+    process = os.popen('wmic memorychip get capacity')
+    result = process.read()
+    process.close()
+    totalMem = 0
+    for m in result.split("  \n\n")[1:-1]:
+        totalMem += int(m)
+    return totalMem / (1024 ** 3)
+
+
+def get_linux_system_memory() -> float:
+    """
+    Obtain the amount of memory available on Linux system.
+    Returns: memory in GB
+    Source: https://stackoverflow.com/a/28161352
+    """
+    import os
+    mem_bytes = os.sysconf('SC_PAGE_SIZE') * os.sysconf('SC_PHYS_PAGES')  # e.g. 4015976448
+    mem_gib = mem_bytes / (1024. ** 3)  # e.g. 3.74
+    return mem_gib
+
+
+def get_mac_system_memory() -> float:
+    """
+    Obtain the amount of memory available on MacOS system.
+    Returns: memory in GB
+    Source: https://apple.stackexchange.com/a/4296
+    """
+
+    import subprocess
+    import re
+
+    # Get process info
+    ps = subprocess.Popen(['ps', '-caxm', '-orss,comm'], stdout=subprocess.PIPE).communicate()[0].decode()
+
+    # Iterate processes
+    processLines = ps.split('\n')
+    sep = re.compile('[\s]+')
+    rssTotal = 0  # kB
+    for row in range(1, len(processLines)):
+        rowText = processLines[row].strip()
+        rowElements = sep.split(rowText)
+        try:
+            rss = float(rowElements[0]) * 1024
+        except:
+            rss = 0  # ignore...
+        rssTotal += rss
+
+    return rssTotal / 1024 ** 3
