@@ -1,10 +1,10 @@
 import logging
 import pytest
-import os
 from testing.functional_tests.t_utils import __tmp_dir__, create_tmp_dir
 from testing.common_testing_util import remove_dataset, remove_tmp_dir
 from ivadomed.scripts import download_data
 from ivadomed.utils import ArgParseException
+from pathlib import Path
 logger = logging.getLogger(__name__)
 
 
@@ -14,10 +14,10 @@ def setup_function():
 
 def test_download_data():
     for dataset in download_data.DICT_URL:
-        output_folder = os.path.join(__tmp_dir__, dataset)
+        output_folder = Path(__tmp_dir__, dataset)
         download_data.main(args=['-d', dataset,
-                                 '-o', output_folder])
-        assert os.path.exists(output_folder)
+                                 '-o', str(output_folder)])
+        assert output_folder.exists()
         remove_dataset(dataset=dataset)
 
 
