@@ -1,9 +1,9 @@
 import logging
-import os
 from testing.functional_tests.t_utils import __tmp_dir__, create_tmp_dir, __data_testing_dir__, \
     download_functional_test_files
 from testing.common_testing_util import remove_tmp_dir
 from ivadomed.scripts import extract_small_dataset
+from pathlib import Path
 logger = logging.getLogger(__name__)
 
 
@@ -12,11 +12,11 @@ def setup_function():
 
 
 def test_extract_small_dataset_default_n(download_functional_test_files):
-    __output_dir__ = os.path.join(__tmp_dir__, 'output_extract_small_dataset')
-    extract_small_dataset.main(args=['--input', __data_testing_dir__,
-                                     '--output', __output_dir__])
-    assert os.path.exists(__output_dir__)
-    output_dir_list = os.listdir(__output_dir__)
+    __output_dir__ = Path(__tmp_dir__, 'output_extract_small_dataset')
+    extract_small_dataset.main(args=['--input', str(__data_testing_dir__),
+                                     '--output', str(__output_dir__)])
+    assert __output_dir__.exists()
+    output_dir_list = [f.name for f in __output_dir__.iterdir()]
     assert 'derivatives' in output_dir_list
     assert 'participants.tsv' in output_dir_list
     assert 'dataset_description.json' in output_dir_list
@@ -26,12 +26,12 @@ def test_extract_small_dataset_default_n(download_functional_test_files):
 
 
 def test_extract_small_dataset_n_2(download_functional_test_files):
-    __output_dir__ = os.path.join(__tmp_dir__, 'output_extract_small_dataset_2')
-    extract_small_dataset.main(args=['--input', __data_testing_dir__,
-                                     '--output', __output_dir__,
+    __output_dir__ = Path(__tmp_dir__, 'output_extract_small_dataset_2')
+    extract_small_dataset.main(args=['--input', str(__data_testing_dir__),
+                                     '--output', str(__output_dir__),
                                      '-n', '2'])
-    assert os.path.exists(__output_dir__)
-    output_dir_list = os.listdir(__output_dir__)
+    assert __output_dir__.exists()
+    output_dir_list = [f.name for f in __output_dir__.iterdir()]
     assert 'derivatives' in output_dir_list
     assert 'participants.tsv' in output_dir_list
     assert 'dataset_description.json' in output_dir_list
@@ -44,12 +44,12 @@ def test_extract_small_dataset_n_2(download_functional_test_files):
 
 
 def test_extract_small_dataset_no_derivatives(download_functional_test_files):
-    __output_dir__ = os.path.join(__tmp_dir__, 'output_extract_small_dataset_3')
-    extract_small_dataset.main(args=['--input', __data_testing_dir__,
-                                     '--output', __output_dir__,
+    __output_dir__ = Path(__tmp_dir__, 'output_extract_small_dataset_3')
+    extract_small_dataset.main(args=['--input', str(__data_testing_dir__),
+                                     '--output', str(__output_dir__),
                                      '-d', '0'])
-    assert os.path.exists(__output_dir__)
-    output_dir_list = os.listdir(__output_dir__)
+    assert __output_dir__.exists()
+    output_dir_list = [f.name for f in __output_dir__.iterdir()]
     assert 'derivatives' not in output_dir_list
     assert 'participants.tsv' in output_dir_list
     assert 'dataset_description.json' in output_dir_list
@@ -59,12 +59,12 @@ def test_extract_small_dataset_no_derivatives(download_functional_test_files):
 
 
 def test_extract_small_dataset_contrast_list(download_functional_test_files):
-    __output_dir__ = os.path.join(__tmp_dir__, 'output_extract_small_dataset_4')
-    extract_small_dataset.main(args=['--input', __data_testing_dir__,
-                                     '--output', __output_dir__,
+    __output_dir__ = Path(__tmp_dir__, 'output_extract_small_dataset_4')
+    extract_small_dataset.main(args=['--input', str(__data_testing_dir__),
+                                     '--output', str(__output_dir__),
                                      '-c', 'T1w, T2w'])
-    assert os.path.exists(__output_dir__)
-    output_dir_list = os.listdir(__output_dir__)
+    assert __output_dir__.exists()
+    output_dir_list = [f.name for f in __output_dir__.iterdir()]
     assert 'derivatives' in output_dir_list
     assert 'participants.tsv' in output_dir_list
     assert 'dataset_description.json' in output_dir_list
