@@ -1,10 +1,10 @@
 import ivadomed.metrics as imed_metrics
 import numpy as np
 import pytest
-import os
 import logging
 from testing.unit_tests.t_utils import create_tmp_dir,  __tmp_dir__
 from testing.common_testing_util import remove_tmp_dir
+from pathlib import Path
 logger = logging.getLogger(__name__)
 
 
@@ -64,17 +64,17 @@ def test_plot_roc_curve():
     tpr = [0, 0.1, 0.5, 0.6, 0.9]
     fpr = [1, 0.8, 0.5, 0.3, 0.1]
     opt_thr_idx = 3
-    __output_file__ = os.path.join(__tmp_dir__, "roc_test.png")
-    imed_metrics.plot_roc_curve(tpr, fpr, opt_thr_idx, __output_file__)
-    assert os.path.isfile(__output_file__)
+    __output_file__ = Path(__tmp_dir__, "roc_test.png")
+    imed_metrics.plot_roc_curve(tpr, fpr, opt_thr_idx, str(__output_file__))
+    assert __output_file__.is_file()
 
 
 def test_dice_plot():
     thr_list = [0.1, 0.3, 0.5, 0.7]
     dice_list = [0.6, 0.7, 0.8, 0.75]
-    __output_file__ = os.path.join(__tmp_dir__, "test_dice.png")
-    imed_metrics.plot_dice_thr(thr_list, dice_list, 2, __output_file__)
-    assert os.path.isfile(__output_file__)
+    __output_file__ = Path(__tmp_dir__, "test_dice.png")
+    imed_metrics.plot_dice_thr(thr_list, dice_list, 2, str(__output_file__))
+    assert __output_file__.is_file()
 
 
 def teardown_function():
