@@ -2,7 +2,9 @@ import pytest
 import numpy as np
 import torch.backends.cudnn as cudnn
 from torch.utils.data import DataLoader
+from loguru import logger
 
+from ivadomed.loader.bids_dataframe import BidsDataframe
 from ivadomed import utils as imed_utils
 from ivadomed.loader import utils as imed_loader_utils, loader as imed_loader
 from testing.unit_tests.t_utils import create_tmp_dir,  __data_testing_dir__, __tmp_dir__, download_data_testing_test_files
@@ -27,7 +29,7 @@ def _cmpt_slice(ds_loader):
                 cmpt_label[1] += 1
             else:
                 cmpt_label[0] += 1
-    print(cmpt_label)
+    logger.debug(cmpt_label)
     return cmpt_label[0], cmpt_label[1]
 
 
@@ -64,20 +66,20 @@ def test_slice_filter(download_data_testing_test_files, transforms_dict, train_l
         "roi_params": roi_params,
         "model_params": {"name": "Unet"},
         "slice_filter_params": slice_filter_params,
+        "patch_filter_params": {"filter_empty_mask": False, "filter_empty_input": False},
         "slice_axis": "axial",
         "multichannel": False
     }
     # Get Training dataset
-    bids_df = imed_loader_utils.BidsDataframe(loader_params, __tmp_dir__, derivatives=True)
+    bids_df = BidsDataframe(loader_params, __tmp_dir__, derivatives=True)
     ds_train = imed_loader.load_dataset(bids_df, **loader_params)
 
-    print('\tNumber of loaded slices: {}'.format(len(ds_train)))
+    logger.info(f"\tNumber of loaded slices: {len(ds_train)}")
 
     train_loader = DataLoader(ds_train, batch_size=BATCH_SIZE,
                               shuffle=True, pin_memory=True,
                               collate_fn=imed_loader_utils.imed_collate,
                               num_workers=0)
-    print('\tNumber of Neg/Pos slices in GT.')
     cmpt_neg, cmpt_pos = _cmpt_slice(train_loader)
     if slice_filter_params["filter_empty_mask"]:
         assert cmpt_neg == 0
@@ -85,6 +87,7 @@ def test_slice_filter(download_data_testing_test_files, transforms_dict, train_l
     else:
         # We verify if there are still some negative slices (they are removed with our filter)
         assert cmpt_neg != 0 and cmpt_pos != 0
+    logger.info(f"\tNumber of Neg/Pos slices in GT: {cmpt_neg/cmpt_pos}")
 
 
 def teardown_function():
