@@ -24,20 +24,16 @@ import nibabel as nib
 import numpy as np
 import seaborn as sns
 from scipy.ndimage import label, generate_binary_structure
+from loguru import logger
 
 from ivadomed import config_manager as imed_config_manager
+from ivadomed import utils as imed_utils
 
 def get_parser():
     parser = argparse.ArgumentParser()
     parser.add_argument("-c", help="Config file path.")
 
     return parser
-
-
-def print_stats(arr):
-    print('\tMean: {}'.format(np.mean(arr)))
-    print('\tMedian: {}'.format(np.median(arr)))
-    print('\tInter-quartile range: [{}, {}]'.format(np.percentile(arr, 25), np.percentile(arr, 75)))
 
 
 def plot_distrib(arr, label, xlim, fname_out):
@@ -52,7 +48,7 @@ def plot_distrib(arr, label, xlim, fname_out):
     plt.xlim(xlim)
     plt.ylabel('Density')
     fig.savefig(fname_out)
-    print('\tSave as: ' + fname_out)
+    logger.info(f"\tSave as: {fname_out}")
 
 
 def run_main(args):
@@ -85,14 +81,14 @@ def run_main(args):
                             vox_lst.append(n_vox)
                             mm3_lst.append(n_vox * px * py * pz)
 
-    print('\nTarget distribution in vox:')
-    print_stats(vox_lst)
+    logger.debug("\nTarget distribution in vox:")
+    imed_utils.print_stats(vox_lst)
     plot_distrib(vox_lst, context["target_suffix"] + ' size in vox',
                  [0, np.percentile(vox_lst, 90)],
                  context["target_suffix"] + '_vox.png')
 
-    print('\nTarget distribution in mm3:')
-    print_stats(mm3_lst)
+    logger.info("\nTarget distribution in mm3:")
+    imed_utils.print_stats(mm3_lst)
     plot_distrib(mm3_lst, context["target_suffix"] + ' size in mm3',
                  [0, np.percentile(mm3_lst, 90)],
                  context["target_suffix"] + '_mm3.png')
