@@ -1,15 +1,23 @@
 import copy
 import random
+from pathlib import Path
+import pickle
+from typing import List, Optional
 
 import numpy as np
-import torch
+
 from torch.utils.data import Dataset
+from loguru import logger
 
 from ivadomed import transforms as imed_transforms, postprocessing as imed_postpro
 from ivadomed.loader import utils as imed_loader_utils
-from ivadomed.loader.utils import dropout_input
+from ivadomed.loader.utils import dropout_input, create_temp_directory, get_obj_size
 from ivadomed.loader.segmentation_pair import SegmentationPair
 from ivadomed.object_detection import utils as imed_obj_detect
+from ivadomed.loader.patch_filter import PatchFilter
+from ivadomed.keywords import MetadataKW, SegmentationDatasetKW, SegmentationPairKW
+from ivadomed.utils import get_timestamp, get_system_memory
+from torchvision.transforms import Compose
 
 
 class MRI3DSubVolumeSegmentationDataset(Dataset):
@@ -24,34 +32,55 @@ class MRI3DSubVolumeSegmentationDataset(Dataset):
 
     Args:
         filename_pairs (list): A list of tuples in the format (input filename, ground truth filename).
-        transform (Compose): Transformations to apply.
         length (tuple): Size of each dimensions of the subvolumes, length equals 3.
         stride (tuple): Size of the overlapping per subvolume and dimensions, length equals 3.
         slice_axis (int): Indicates the axis used to extract slices: "axial": 2, "sagittal": 0, "coronal": 1.
+        transform (Compose): Transformations to apply.
+        subvolume_filter_fn (PatchFilter): PatchFilter object containing subvolume filter parameters.
         soft_gt (bool): If True, ground truths are not binarized before being fed to the network. Otherwise, ground
         truths are thresholded (0.5) after the data augmentation operations.
         is_input_dropout (bool): Return input with missing modalities.
+        disk_cache (bool): set whether all input data should be cached in local folders to allow faster subsequent
+        reloading and bypass memory cap.
     """
 
-    def __init__(self, filename_pairs, transform=None, length=(64, 64, 64), stride=(0, 0, 0), slice_axis=0,
-                 task="segmentation", soft_gt=False, is_input_dropout=False):
+    def __init__(self,
+                 filename_pairs: list,
+                 transform: List[Optional[Compose]] = None,
+                 length: tuple = (64, 64, 64),
+                 stride: tuple = (0, 0, 0),
+                 slice_axis: int = 0,
+                 subvolume_filter_fn: PatchFilter = None,
+                 task: str = "segmentation",
+                 soft_gt: bool = False,
+                 is_input_dropout: bool = False,
+                 disk_cache: bool=True):
+
         self.filename_pairs = filename_pairs
-        self.handlers = []
-        self.indexes = []
+
+        # could be a list of tuple of objects OR path objects to the actual disk equivalent.
+        # behaves differently depend on if self.cache is set to true or not.
+        self.handlers: List[tuple] = []
+
+        self.indexes: list = []
         self.length = length
         self.stride = stride
         self.prepro_transforms, self.transform = transform
         self.slice_axis = slice_axis
-        self.has_bounding_box = True
+        self.subvolume_filter_fn = subvolume_filter_fn
+        self.has_bounding_box: bool = True
         self.task = task
         self.soft_gt = soft_gt
         self.is_input_dropout = is_input_dropout
+        self.disk_cache: bool = disk_cache
 
         self._load_filenames()
         self._prepare_indices()
 
-    def _load_filenames(self):
+    def _load_filenames(self) -> None:
         """Load preprocessed pair data (input and gt) in handler."""
+        path_temp: Path = Path(create_temp_directory())
+
         for input_filename, gt_filename, roi_filename, metadata in self.filename_pairs:
             segpair = SegmentationPair(input_filename, gt_filename, metadata=metadata, slice_axis=self.slice_axis,
                                        soft_gt=self.soft_gt)
@@ -60,27 +89,54 @@ class MRI3DSubVolumeSegmentationDataset(Dataset):
             seg_pair = {
                 'input': input_data,
                 'gt': gt_data,
-                'input_metadata': metadata['input_metadata'],
-                'gt_metadata': metadata['gt_metadata']
+                MetadataKW.INPUT_METADATA: metadata[MetadataKW.INPUT_METADATA],
+                MetadataKW.GT_METADATA: metadata[MetadataKW.GT_METADATA]
             }
 
             self.has_bounding_box = imed_obj_detect.verify_metadata(seg_pair, self.has_bounding_box)
             if self.has_bounding_box:
-                self.prepro_transforms = imed_obj_detect.adjust_transforms(self.prepro_transforms, seg_pair,
+                self.prepro_transforms = imed_obj_detect.adjust_transforms(self.prepro_transforms,
+                                                                           seg_pair,
                                                                            length=self.length,
                                                                            stride=self.stride)
+
             seg_pair, roi_pair = imed_transforms.apply_preprocessing_transforms(self.prepro_transforms,
                                                                                 seg_pair=seg_pair)
 
-            for metadata in seg_pair['input_metadata']:
-                metadata['index_shape'] = seg_pair['input'][0].shape
-            self.handlers.append((seg_pair, roi_pair))
+            for metadata in seg_pair[MetadataKW.INPUT_METADATA]:
+                metadata[MetadataKW.INDEX_SHAPE] = seg_pair['input'][0].shape
+
+            # First time detemine cache automatically IF not specified. Otherwise, use the cache specified.
+            if self.disk_cache is None:
+                self.disk_cache = self.determine_cache_need(seg_pair, roi_pair)
+
+            if self.disk_cache:
+                # Write SegPair and ROIPair to disk cache with timestamp to avoid collisions
+                # 'self.handler' is now a list of a FILES instead of actual data to prevent using too much memory
+                path_cache_seg_pair = path_temp / f'seg_pair_{get_timestamp()}.pkl'
+                with path_cache_seg_pair.open(mode='wb') as f:
+                    pickle.dump(seg_pair, f)
+
+                path_cache_roi_pair = path_temp / f'roi_pair_{get_timestamp()}.pkl'
+                with path_cache_roi_pair.open(mode='wb') as f:
+                    pickle.dump(roi_pair, f)
+                self.handlers.append((path_cache_seg_pair, path_cache_roi_pair))
+
+            else:
+                self.handlers.append((seg_pair, roi_pair))
 
     def _prepare_indices(self):
         """Stores coordinates of subvolumes for training."""
         for i in range(0, len(self.handlers)):
-            segpair, _ = self.handlers[i]
-            input_img = self.handlers[i][0]['input']
+
+            if self.disk_cache:
+                with self.handlers[i][0].open(mode='rb') as f:
+                    segpair = pickle.load(f)
+            else:
+                segpair = self.handlers[i][0]
+
+            input_img, gt = segpair.get('input'), segpair.get('gt')
+
             shape = input_img[0].shape
 
             if ((shape[0] - self.length[0]) % self.stride[0]) != 0 or self.length[0] % 16 != 0 or shape[0] < \
@@ -95,91 +151,163 @@ class MRI3DSubVolumeSegmentationDataset(Dataset):
             for x in range(0, (shape[0] - self.length[0]) + 1, self.stride[0]):
                 for y in range(0, (shape[1] - self.length[1]) + 1, self.stride[1]):
                     for z in range(0, (shape[2] - self.length[2]) + 1, self.stride[2]):
-                        self.indexes.append({
-                            'x_min': x,
-                            'x_max': x + self.length[0],
-                            'y_min': y,
-                            'y_max': y + self.length[1],
-                            'z_min': z,
-                            'z_max': z + self.length[2],
-                            'handler_index': i})
+                        x_min, x_max = x, x + self.length[0]
+                        y_min, y_max = y, y + self.length[1]
+                        z_min, z_max = z, z + self.length[2]
 
-    def __len__(self):
+                        subvolume = {
+                            'input': list(np.asarray(input_img)[:, x_min:x_max, y_min:y_max, z_min:z_max]),
+                            'gt': list(np.asarray(gt)[:, x_min:x_max, y_min:y_max, z_min:z_max] if gt else []),
+                        }
+
+                        if self.subvolume_filter_fn and not self.subvolume_filter_fn(subvolume):
+                            continue
+
+                        self.indexes.append({
+                            'x_min': x_min, 
+                            'x_max': x_max, 
+                            'y_min': y_min,
+                            'y_max': y_max, 
+                            'z_min': z_min,
+                            'z_max': z_max,
+                            'handler_index': i,
+                        })
+
+    def __len__(self) -> int:
         """Return the dataset size. The number of subvolumes."""
         return len(self.indexes)
 
-    def __getitem__(self, index):
-        """Return the specific index pair subvolume (input, ground truth).
+    def __getitem__(self, subvolume_index: int) -> dict:
+        """Return the specific processed subvolume corresponding to index (input, ground truth and metadata).
 
         Args:
-            index (int): Subvolume index.
+            subvolume_index (int): Subvolume (patch) index.
         """
 
+        # CONTEXT
+        # All 3D models are trained with 3D subvolumes (patches):
+        #    * 'self.handlers' contains paired data for all preprocessed 3D volumes
+        #    * 'self.indexes' is a list of coordinates for all 3D subvolumes
+        #      e.g. [{'x_min': 0, 'x_max': 32, 'y_min': 0, 'y_max': 32, 'z_min': 0, 'z_max': 16, 'handler_index': 0},
+        #            {'x_min': 0, 'x_max': 32, 'y_min': 0, 'y_max': 32, 'z_min': 16, 'z_max': 32, 'handler_index': 0}]
+        #      where 'handler_index' is the index of the 3D volume from which the subvolume is extracted
+        # Note that ROI is not available for 3D models
+
+        # Extract coordinates and paired data for the subvolume
+        # Get subvolume coordinates from 'self.indexes'
+        coord: dict = self.indexes[subvolume_index]
+        # Extract subvolume pair from 'self.handlers'
         # copy.deepcopy is used to have different coordinates for reconstruction for a given handler,
         # to allow a different rater at each iteration of training, and to clean transforms params from previous
         # transforms i.e. remove params from previous iterations so that the coming transforms are different
-        coord = self.indexes[index]
-        seg_pair, _ = copy.deepcopy(self.handlers[coord['handler_index']])
+        tuple_seg_roi_pair: tuple = self.handlers[coord.get(SegmentationDatasetKW.HANDLER_INDEX)]
+        if self.disk_cache:
+            with tuple_seg_roi_pair[0].open(mode='rb') as f:
+                seg_pair = pickle.load(f)
+        else:
+            seg_pair, _ = copy.deepcopy(tuple_seg_roi_pair)
 
-        # In case multiple raters
-        if seg_pair['gt'] and isinstance(seg_pair['gt'][0], list):
+        # In case of multiple raters
+        if seg_pair[SegmentationPairKW.GT] and isinstance(seg_pair[SegmentationPairKW.GT][0], list):
             # Randomly pick a rater
-            idx_rater = random.randint(0, len(seg_pair['gt'][0]) - 1)
+            idx_rater = random.randint(0, len(seg_pair[SegmentationPairKW.GT][0]) - 1)
             # Use it as ground truth for this iteration
             # Note: in case of multi-class: the same rater is used across classes
-            for idx_class in range(len(seg_pair['gt'])):
-                seg_pair['gt'][idx_class] = seg_pair['gt'][idx_class][idx_rater]
-                seg_pair['gt_metadata'][idx_class] = seg_pair['gt_metadata'][idx_class][idx_rater]
+            for idx_class in range(len(seg_pair[SegmentationPairKW.GT])):
+                seg_pair[SegmentationPairKW.GT][idx_class] = seg_pair[SegmentationPairKW.GT][idx_class][idx_rater]
+                seg_pair[SegmentationPairKW.GT_METADATA][idx_class] = seg_pair[SegmentationPairKW.GT_METADATA][idx_class][idx_rater]
 
-        metadata_input = seg_pair['input_metadata'] if seg_pair['input_metadata'] is not None else []
-        metadata_gt = seg_pair['gt_metadata'] if seg_pair['gt_metadata'] is not None else []
+        # Extract metadata from paired data
+        metadata_input = seg_pair[SegmentationPairKW.INPUT_METADATA] if seg_pair[SegmentationPairKW.INPUT_METADATA] is not None else []
+        metadata_gt = seg_pair[SegmentationPairKW.GT_METADATA] if seg_pair[SegmentationPairKW.GT_METADATA] is not None else []
 
-        # Run transforms on images
-        stack_input, metadata_input = self.transform(sample=seg_pair['input'],
+        # Extract min/max coordinates
+        x_min = coord.get(SegmentationDatasetKW.X_MIN)
+        x_max = coord.get(SegmentationDatasetKW.X_MAX)
+        y_min = coord.get(SegmentationDatasetKW.Y_MIN)
+        y_max = coord.get(SegmentationDatasetKW.Y_MAX)
+        z_min = coord.get(SegmentationDatasetKW.Z_MIN)
+        z_max = coord.get(SegmentationDatasetKW.Z_MAX)
+
+        # Extract subvolume and gt from coordinates
+        stack_input = np.asarray(seg_pair[SegmentationPairKW.INPUT])[
+                      :,
+                      x_min:x_max,
+                      y_min:y_max,
+                      z_min:z_max
+        ]
+
+        if seg_pair[SegmentationPairKW.GT]:
+            stack_gt = np.asarray(seg_pair[SegmentationPairKW.GT])[
+                       :,
+                       x_min:x_max,
+                       y_min:y_max,
+                       z_min:z_max
+            ]
+        else:
+            stack_gt = []
+
+        # Run transforms on subvolume
+        stack_input, metadata_input = self.transform(sample=list(stack_input),
                                                      metadata=metadata_input,
                                                      data_type="im")
         # Update metadata_gt with metadata_input
         metadata_gt = imed_loader_utils.update_metadata(metadata_input, metadata_gt)
 
-        # Run transforms on images
-        stack_gt, metadata_gt = self.transform(sample=seg_pair['gt'],
+        # Run transforms on gt
+        stack_gt, metadata_gt = self.transform(sample=list(stack_gt),
                                                metadata=metadata_gt,
                                                data_type="gt")
         # Make sure stack_gt is binarized
         if stack_gt is not None and not self.soft_gt:
             stack_gt = imed_postpro.threshold_predictions(stack_gt, thr=0.5).astype(np.uint8)
 
-        shape_x = coord["x_max"] - coord["x_min"]
-        shape_y = coord["y_max"] - coord["y_min"]
-        shape_z = coord["z_max"] - coord["z_min"]
-
-        # add coordinates to metadata to reconstruct volume
+        # Add coordinates to metadata to reconstruct volume
         for metadata in metadata_input:
-            metadata['coord'] = [coord["x_min"], coord["x_max"], coord["y_min"], coord["y_max"], coord["z_min"],
-                                 coord["z_max"]]
+            metadata[MetadataKW.COORD] = [
+                x_min, x_max,
+                y_min, y_max,
+                z_min, z_max,
+            ]
 
+        # Combine all processed data for a given subvolume in dictionary
         subvolumes = {
-            'input': torch.zeros(stack_input.shape[0], shape_x, shape_y, shape_z),
-            'gt': torch.zeros(stack_gt.shape[0], shape_x, shape_y, shape_z) if stack_gt is not None else None,
-            'input_metadata': metadata_input,
-            'gt_metadata': metadata_gt
+            SegmentationPairKW.INPUT: stack_input,
+            SegmentationPairKW.GT: stack_gt,
+            MetadataKW.INPUT_METADATA: metadata_input,
+            MetadataKW.GT_METADATA: metadata_gt
         }
-
-        for _ in range(len(stack_input)):
-            subvolumes['input'] = stack_input[:,
-                                  coord['x_min']:coord['x_max'],
-                                  coord['y_min']:coord['y_max'],
-                                  coord['z_min']:coord['z_max']]
 
         # Input-level dropout to train with missing modalities
         if self.is_input_dropout:
             subvolumes = dropout_input(subvolumes)
 
-        if stack_gt is not None:
-            for _ in range(len(stack_gt)):
-                subvolumes['gt'] = stack_gt[:,
-                                   coord['x_min']:coord['x_max'],
-                                   coord['y_min']:coord['y_max'],
-                                   coord['z_min']:coord['z_max']]
-
         return subvolumes
+
+    def determine_cache_need(self, seg_pair: dict, roi_pair: dict):
+        """
+        When Cache flag is not explicitly set, determine whether to cache the data or not
+        Args:
+            seg_pair: an EXAMPLE, typical seg_pair object
+            roi_pair: an EXAMPLE, typical seg_pair object
+
+        Returns:
+
+        """
+        size_seg_pair_in_bytes = get_obj_size(seg_pair)
+        size_roi_pair_in_bytes = get_obj_size(roi_pair)
+
+        optimal_ram_limit = get_system_memory() * 0.5
+
+        # Size limit: 4GB GPU RAM, keep in mind tranform etc might take MORE!
+        size_estimated_dataset_GB = (size_seg_pair_in_bytes + size_roi_pair_in_bytes) * len(self.filename_pairs) / 1024 ** 3
+        if size_estimated_dataset_GB > optimal_ram_limit:
+            logger.info(f"Estimated 3D dataset size is {size_estimated_dataset_GB} GB, which is larger than {optimal_ram_limit} GB. Auto "
+                        f"enabling cache.")
+            self.disk_cache = True
+            return True
+        else:
+            logger.info(f"Estimated 3D dataset size is {size_estimated_dataset_GB} GB, which is smaller than {optimal_ram_limit} GB. File "
+                        f"cache will not be used")
+            self.disk_cache = False
+            return False
